Recreate a BaseMesh Implemention defining the structure of what makes a discretize mesh
