+++ discretize/base/__init__.py
@@ -12,9 +12,10 @@ Base Mesh Class
   :toctree: generated/
 
   BaseMesh
+  BaseRegularMesh
   BaseRectangularMesh
   BaseTensorMesh
 """
-
-from discretize.base.base_regular_mesh import BaseMesh, BaseRectangularMesh
+from discretize.base.base_mesh import BaseMesh
+from discretize.base.base_regular_mesh import BaseRegularMesh, BaseRectangularMesh
 from discretize.base.base_tensor_mesh import BaseTensorMesh
