+++ discretize/curvilinear_mesh.py
@@ -10,7 +10,6 @@ from discretize.utils import (
 from discretize.base import BaseRectangularMesh
 from discretize.operators import DiffOperators, InnerProducts
 from discretize.mixins import InterfaceMixins
-from discretize.utils.code_utils import deprecate_property
 
 # Some helper functions.
 def _length2D(x):
@@ -82,8 +81,6 @@ class CurvilinearMesh(
         **DiffOperators._aliases,
         **BaseRectangularMesh._aliases,
         **{
-            "gridCC": "cell_centers",
-            "gridN": "nodes",
             "gridFx": "faces_x",
             "gridFy": "faces_y",
             "gridFz": "faces_z",
@@ -150,18 +147,6 @@ class CurvilinearMesh(
 
     @property
     def cell_centers(self):
-        """Gridded cell center locations (staggered grid)
-
-        For 2D or 3D curvilinear meshes, this property returns a numpy array
-        of shape (n_cells, dim) containing the gridded cell center locations.
-        The bottom-front-leftmost cell is the first cell.
-
-        Returns
-        -------
-        (n_cells, dim) numpy.ndarray of float
-            The shape of the output array is the number of cells by the dimension.
-
-        """
         if getattr(self, "_cell_centers", None) is None:
             self._cell_centers = np.concatenate(
                 [self.aveN2CC * self.gridN[:, i] for i in range(self.dim)]
@@ -170,17 +155,6 @@ class CurvilinearMesh(
 
     @property
     def nodes(self):
-        """Gridded node locations (staggered grid)
-
-        For 2D or 3D curvilinear meshes, this property returns a numpy array
-        of shape (n_nodes, dim) containing the gridded node locations.
-        The bottom-front-leftmost node is the first node.
-
-        Returns
-        -------
-        (n_nodes, dim) numpy.ndarray of float
-            Gridded node locations
-        """
         if getattr(self, "_nodes", None) is None:
             raise Exception("Someone deleted this. I blame you.")
         return self._nodes
@@ -364,60 +338,6 @@ class CurvilinearMesh(
 
     @property
     def faces(self):
-        """Gridded face locations (staggered grid)
-
-        This property returns a numpy array of shape (n_faces, dim)
-        containing gridded locations for all faces in the
-        mesh (staggered grid). This is equivalent to calling
-        np.r_[faces_x, faces_y, faces_z].
-
-        Returns
-        -------
-        (n_faces, dim) numpy.ndarray of float
-            Gridded face locations (staggered grid)
-
-        Examples
-        --------
-        Here, we provide an example of a minimally staggered curvilinear mesh.
-        In this case, the x and y-faces have normal vectors that are
-        primarily along the x and y-directions, respectively.
-
-        >>> from discretize import CurvilinearMesh
-        >>> from discretize.utils import example_curvilinear_grid, mkvc
-        >>> from matplotlib import pyplot as plt
-
-        >>> x, y = example_curvilinear_grid([10, 10], "rotate")
-        >>> mesh1 = CurvilinearMesh([x, y])
-        >>> faces = mesh1.faces
-        >>> x_faces = faces[:mesh1.n_faces_x]
-        >>> y_faces = faces[mesh1.n_faces_x:]
-
-        >>> fig1 = plt.figure(figsize=(5, 5))
-        >>> ax1 = fig1.add_subplot(111)
-        >>> mesh1.plot_grid(ax=ax1)
-        >>> ax1.scatter(x_faces[:, 0], x_faces[:, 1], 30, 'r')
-        >>> ax1.scatter(y_faces[:, 0], y_faces[:, 1], 30, 'g')
-        >>> ax1.legend(['Mesh', 'X-faces', 'Y-faces'], fontsize=16)
-        >>> plt.plot()
-
-        Here, we provide an example of a highly irregular curvilinear mesh.
-        In this case, the y-faces are not defined by normal vectors along
-        a particular direction.
-
-        >>> x, y = example_curvilinear_grid([10, 10], "sphere")
-        >>> mesh2 = CurvilinearMesh([x, y])
-        >>> faces = mesh2.faces
-        >>> x_faces = faces[:mesh2.n_faces_x]
-        >>> y_faces = faces[mesh2.n_faces_x:]
-
-        >>> fig2 = plt.figure(figsize=(5, 5))
-        >>> ax2 = fig2.add_subplot(111)
-        >>> mesh2.plot_grid(ax=ax2)
-        >>> ax2.scatter(x_faces[:, 0], x_faces[:, 1], 30, 'r')
-        >>> ax2.scatter(y_faces[:, 0], y_faces[:, 1], 30, 'g')
-        >>> ax2.legend(['Mesh', 'X-faces', 'Y-faces'], fontsize=16)
-        >>> plt.plot()
-        """
         faces = np.r_[self.faces_x, self.faces_y]
         if self.dim > 2:
             faces = np.r_[faces, self.faces_z]
@@ -569,60 +489,6 @@ class CurvilinearMesh(
 
     @property
     def edges(self):
-        """Gridded edge locations (staggered grid)
-
-        This property returns a numpy array of shape (n_edges, dim)
-        containing gridded locations for all edges in the
-        mesh (staggered grid). This is equivalent to calling
-        np.r_[edges_x, edges_y, edges_z].
-
-        Returns
-        -------
-        (n_edges, dim) numpy.ndarray of float
-            Gridded edge locations (staggered grid)
-
-        Examples
-        --------
-        Here, we provide an example of a minimally staggered curvilinear mesh.
-        In this case, the x and y-edges have normal vectors that are
-        primarily along the x and y-directions, respectively.
-
-        >>> from discretize import CurvilinearMesh
-        >>> from discretize.utils import example_curvilinear_grid, mkvc
-        >>> from matplotlib import pyplot as plt
-
-        >>> x, y = example_curvilinear_grid([10, 10], "rotate")
-        >>> mesh1 = CurvilinearMesh([x, y])
-        >>> edges = mesh1.edges
-        >>> x_edges = edges[:mesh1.n_edges_x]
-        >>> y_edges = edges[mesh1.n_edges_x:]
-
-        >>> fig1 = plt.figure(figsize=(5, 5))
-        >>> ax1 = fig1.add_subplot(111)
-        >>> mesh1.plot_grid(ax=ax1)
-        >>> ax1.scatter(x_edges[:, 0], x_edges[:, 1], 30, 'r')
-        >>> ax1.scatter(y_edges[:, 0], y_edges[:, 1], 30, 'g')
-        >>> ax1.legend(['Mesh', 'X-edges', 'Y-edges'], fontsize=16)
-        >>> plt.plot()
-
-        Here, we provide an example of a highly irregular curvilinear mesh.
-        In this case, the y-edges are not defined by normal vectors along
-        a particular direction.
-
-        >>> x, y = example_curvilinear_grid([10, 10], "sphere")
-        >>> mesh2 = CurvilinearMesh([x, y])
-        >>> edges = mesh2.edges
-        >>> x_edges = edges[:mesh2.n_edges_x]
-        >>> y_edges = edges[mesh2.n_edges_x:]
-
-        >>> fig2 = plt.figure(figsize=(5, 5))
-        >>> ax2 = fig2.add_subplot(111)
-        >>> mesh2.plot_grid(ax=ax2)
-        >>> ax2.scatter(x_edges[:, 0], x_edges[:, 1], 30, 'r')
-        >>> ax2.scatter(y_edges[:, 0], y_edges[:, 1], 30, 'g')
-        >>> ax2.legend(['Mesh', 'X-edges', 'Y-edges'], fontsize=16)
-        >>> plt.show()
-        """
         edges = np.r_[self.edges_x, self.edges_y]
         if self.dim > 2:
             edges = np.r_[edges, self.edges_z]
@@ -630,33 +496,10 @@ class CurvilinearMesh(
 
     @property
     def boundary_nodes(self):
-        """Gridded boundary node locations
-
-        This property returns a numpy array of shape
-        (n_boundary_nodes, dim) containing the gridded locations
-        of the nodes on the boundary of the mesh.
-
-        Returns
-        -------
-        (n_boundary_nodes, dim) numpy.ndarray of float
-            Gridded boundary node locations
-        """
         return self.nodes[make_boundary_bool(self.shape_nodes)]
 
     @property
     def boundary_edges(self):
-        """Gridded boundary edge locations
-
-        This property returns a numpy array of shape
-        (n_boundary_edges, dim) containing the gridded locations
-        of the edges on the boundary of the mesh. The returned
-        quantity is organized *np.r_[edges_x, edges_y, edges_z]* .
-
-        Returns
-        -------
-        (n_boundary_edges, dim) numpy.ndarray of float
-            Gridded boundary edge locations
-        """
         if self.dim == 2:
             ex = self.edges_x[make_boundary_bool(self.shape_edges_x, dir="y")]
             ey = self.edges_y[make_boundary_bool(self.shape_edges_y, dir="x")]
@@ -669,16 +512,6 @@ class CurvilinearMesh(
 
     @property
     def boundary_faces(self):
-        """Gridded locations of non-hanging x-faces
-
-        This property returns a numpy array of shape (n_faces_x, dim)
-        containing gridded locations for all non-hanging x-faces.
-
-        Returns
-        -------
-        (n_faces_x, dim) numpy.ndarray of float
-            Gridded locations of all non-hanging x-faces
-        """
         fx = self.faces_x[make_boundary_bool(self.shape_faces_x, dir="x")]
         fy = self.faces_y[make_boundary_bool(self.shape_faces_y, dir="y")]
         if self.dim == 2:
@@ -689,18 +522,6 @@ class CurvilinearMesh(
 
     @property
     def boundary_face_outward_normals(self):
-        """Outward normals of boundary faces
-
-        For all boundary faces in the mesh, this property returns
-        the unit vectors denoting the outward normals to the boundary.
-        The returned quantity is a numpy array of shape
-        (n_boundary_faces, dim).
-
-        Returns
-        -------
-        (n_boundary_faces, dim) numpy.ndarray of float
-            Outward normals of boundary faces
-        """
         is_bxm = np.zeros(self.shape_faces_x, order="F", dtype=bool)
         is_bxm[0, :] = True
         is_bxm = is_bxm.reshape(-1, order="F")
@@ -758,19 +579,6 @@ class CurvilinearMesh(
 
     @property
     def cell_volumes(self):
-        """Return cell volumes
-
-        Calling this property will compute and return a 1D array
-        containing the volumes of mesh cells.
-
-        Returns
-        -------
-        (n_cells) numpy.ndarray
-            The quantity returned depends on the dimensions of the mesh:
-                - *2D:* Returns the cell areas
-                - *3D:* Returns the cell volumes
-        """
-
         if getattr(self, "_cell_volumes", None) is None:
             if self.dim == 2:
                 A, B, C, D = index_cube("ABCD", self.vnN)
@@ -805,22 +613,6 @@ class CurvilinearMesh(
 
     @property
     def face_areas(self):
-        """Returns the areas of all faces in the mesh
-
-        Calling this property will compute and return the areas of all
-        faces as a 1D numpy array. The returned quantity is ordered x-face
-        areas, then y-face areas, then z-face areas.
-
-        Returns
-        -------
-        (n_faces) numpy.ndarray
-            The length of the quantity returned depends on the dimensions of the mesh:
-
-            - *1D:* returns the x-face areas
-            - *2D:* returns the x-face and y-face areas in order; i.e. y-edge
-              and x-edge lengths, respectively
-            - *3D:* returns the x, y and z-face areas in order
-        """
         if (
             getattr(self, "_face_areas", None) is None
             or getattr(self, "_normals", None) is None
@@ -866,21 +658,10 @@ class CurvilinearMesh(
 
     @property
     def face_normals(self):
-        """Gridded average face normals for all mesh faces.
-
-        This property computes and returns a numpy array of shape
-        (n_faces, dim) containing the normal vectors for
-        all mesh faces. For 3D meshes, there are 4 nodes in which
-        cross-products can be used to compute the normal vector.
-        In this case, the average normal vector is returned so there
-        is only 1 vector per face.
-
-        Returns
-        -------
-        (n_faces, dim) numpy.ndarray of shape
-            Gridded average face normals for all mesh faces.
-        """
-
+        # For 3D meshes, there are 4 nodes in which
+        # cross-products can be used to compute the normal vector.
+        # In this case, the average normal vector is returned so there
+        # is only 1 vector per face.
         if getattr(self, "_normals", None) is None:
             self.face_areas  # calling .face_areas will create the face normals
         if self.dim == 2:
@@ -908,21 +689,6 @@ class CurvilinearMesh(
 
     @property
     def edge_lengths(self):
-        """Returns the lengths of all edges in the mesh
-
-        Calling this property will compute and return the lengths of all
-        edges in the mesh. The returned quantity is ordered x-edge lengths,
-        then y-edge lengths, then z-edge lengths.
-
-        Returns
-        -------
-        (n_edges) numpy.ndarray
-            The length of the quantity returned depends on the dimensions of the mesh:
-
-            - *1D:* returns the x-edge lengths
-            - *2D:* returns the x-edge and y-edge lengths in order
-            - *3D:* returns the x, y and z-edge lengths in order
-        """
         if getattr(self, "_edge_lengths", None) is None:
             if self.dim == 2:
                 xy = self.gridN
@@ -957,24 +723,6 @@ class CurvilinearMesh(
 
     @property
     def edge_tangents(self):
-        """Gridded edge tangents directions for all mesh edges
-
-        This property computes and returns a numpy array of shape
-        (n_edges, dim) containing the edge tangent directions for
-        all mesh edges.
-
-        Returns
-        -------
-        (n_edges, dim) numpy.ndarray of shape
-            Gridded edge tangent directions for all mesh edges
-        """
         if getattr(self, "_edge_tangents", None) is None:
             self.edge_lengths  # calling .edge_lengths will create the tangents
         return self._edge_tangents
-
-    # DEPRECATIONS
-    vol = deprecate_property("cell_volumes", "vol", removal_version="1.0.0", future_warn=False)
-    area = deprecate_property("face_areas", "area", removal_version="1.0.0", future_warn=False)
-    edge = deprecate_property("edge_lengths", "edge", removal_version="1.0.0", future_warn=False)
-    # tangent already deprecated in BaseMesh
-    # normals already deprecated in BaseMesh
