+++ discretize/cylindrical_mesh.py
@@ -1869,15 +1869,12 @@ class CylindricalMesh(
         return Proj * Pc2r
 
     # DEPRECATIONS
-    vol = deprecate_property("cell_volumes", "vol", removal_version="1.0.0", future_warn=False)
-    area = deprecate_property("face_areas", "area", removal_version="1.0.0", future_warn=False)
     areaFx = deprecate_property("face_x_areas", "areaFx", removal_version="1.0.0", future_warn=False)
     areaFy = deprecate_property("face_y_areas", "areaFy", removal_version="1.0.0", future_warn=False)
     areaFz = deprecate_property("face_z_areas", "areaFz", removal_version="1.0.0", future_warn=False)
     edgeEx = deprecate_property("edge_x_lengths", "edgeEx", removal_version="1.0.0", future_warn=False)
     edgeEy = deprecate_property("edge_y_lengths", "edgeEy", removal_version="1.0.0", future_warn=False)
     edgeEz = deprecate_property("edge_z_lengths", "edgeEz", removal_version="1.0.0", future_warn=False)
-    edge = deprecate_property("edge_lengths", "edge", removal_version="1.0.0", future_warn=False)
     isSymmetric = deprecate_property(
         "is_symmetric", "isSymmetric", removal_version="1.0.0", future_warn=False
     )
