+++ discretize/operators/differential_operators.py
@@ -145,21 +145,12 @@ class DiffOperators(object):
     """
 
     _aliases = {
-        "aveF2CC": "average_face_to_cell",
-        "aveF2CCV": "average_face_to_cell_vector",
         "aveFx2CC": "average_face_x_to_cell",
         "aveFy2CC": "average_face_y_to_cell",
         "aveFz2CC": "average_face_z_to_cell",
-        "aveCC2F": "average_cell_to_face",
-        "aveCCV2F": "average_cell_vector_to_face",
-        "aveE2CC": "average_edge_to_cell",
-        "aveE2CCV": "average_edge_to_cell_vector",
         "aveEx2CC": "average_edge_x_to_cell",
         "aveEy2CC": "average_edge_y_to_cell",
         "aveEz2CC": "average_edge_z_to_cell",
-        "aveN2CC": "average_node_to_cell",
-        "aveN2E": "average_node_to_edge",
-        "aveN2F": "average_node_to_face",
     }
 
     ###########################################################################
@@ -241,154 +232,6 @@ class DiffOperators(object):
 
     @property
     def face_divergence(self):
-        """Face divergence operator (faces to cell-centres)
-
-        This property constructs the 2nd order numerical divergence operator
-        that maps from faces to cell centers. The operator is a sparse matrix
-        :math:`\\mathbf{D_f}` that can be applied as a matrix-vector product to
-        a discrete vector :math:`\\mathbf{u}` that lives on mesh faces; i.e.::
-
-            div_u = Df @ u
-
-        Once constructed, the operator is stored permanently as a property of the mesh.
-        *See notes for additional details.*
-
-        Returns
-        -------
-        (n_cells, n_faces) scipy.sparse.csr_matrix
-            The numerical divergence operator from faces to cell centers
-
-        Notes
-        -----
-        In continuous space, the divergence operator is defined as:
-
-        .. math::
-            \\phi = \\nabla \\cdot \\vec{u} = \\frac{\\partial u_x}{\\partial x}
-            + \\frac{\\partial u_y}{\\partial y} + \\frac{\\partial u_z}{\\partial z}
-
-        Where :math:`\\mathbf{u}` is the discrete representation of the continuous variable
-        :math:`\\vec{u}` on cell faces and :math:`\\boldsymbol{\\phi}` is the discrete
-        representation of :math:`\\phi` at cell centers, **face_divergence** constructs a
-        discrete linear operator :math:`\\mathbf{D_f}` such that:
-
-        .. math::
-            \\boldsymbol{\\phi} = \\mathbf{D_f \\, u}
-
-        For each cell, the computation of the face divergence can be expressed
-        according to the integral form below. For cell :math:`i` whose corresponding
-        faces are indexed as a subset :math:`K` from the set of all mesh faces:
-
-        .. math::
-            \\phi_i = \\frac{1}{V_i} \\sum_{k \\in K} A_k \\, \\vec{u}_k \\cdot \\hat{n}_k
-
-        where :math:`V_i` is the volume of cell :math:`i`, :math:`A_k` is
-        the surface area of face *k*, :math:`\\vec{u}_k` is the value of
-        :math:`\\vec{u}` on face *k*, and :math:`\\hat{n}_k`
-        represents the outward normal vector of face *k* for cell *i*.
-
-
-        Examples
-        --------
-        Below, we demonstrate 1) how to apply the face divergence operator to
-        a discrete vector and 2) the mapping of the face divergence operator and
-        its sparsity. Our example is carried out on a 2D mesh but it can
-        be done equivalently for a 3D mesh.
-
-        We start by importing the necessary packages and modules.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> import matplotlib as mpl
-
-        Define a 2D mesh
-
-        >>> h = np.ones(20)
-        >>> mesh = TensorMesh([h, h], "CC")
-
-        Create a discrete vector on mesh faces
-
-        >>> faces_x = mesh.faces_x
-        >>> faces_y = mesh.faces_y
-        >>> ux = (faces_x[:, 0] / np.sqrt(np.sum(faces_x ** 2, axis=1))) * np.exp(
-        ...     -(faces_x[:, 0] ** 2 + faces_x[:, 1] ** 2) / 6 ** 2
-        ... )
-        >>> uy = (faces_y[:, 1] / np.sqrt(np.sum(faces_y ** 2, axis=1))) * np.exp(
-        ...     -(faces_y[:, 0] ** 2 + faces_y[:, 1] ** 2) / 6 ** 2
-        ... )
-        >>> u = np.r_[ux, uy]
-
-        Construct the divergence operator and apply to face-vector
-
-        >>> Df = mesh.face_divergence
-        >>> div_u = Df @ u
-
-        Plot the original face-vector and its divergence
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(13, 6))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(
-            ...     u, ax=ax1, v_type="F", view="vec", stream_opts={"color": "w", "density": 1.0}
-            ... )
-            >>> ax1.set_title("Vector at cell faces", fontsize=14)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(div_u, ax=ax2)
-            >>> ax2.set_yticks([])
-            >>> ax2.set_ylabel("")
-            >>> ax2.set_title("Divergence at cell centers", fontsize=14)
-            >>> plt.show()
-
-        The discrete divergence operator is a sparse matrix that maps
-        from faces to cell centers. To demonstrate this, we construct
-        a small 2D mesh. We then show the ordering of the elements in
-        the original discrete quantity :math:`\\mathbf{u}` and its
-        discrete divergence :math:`\\boldsymbol{\\phi}` as well as a
-        spy plot.
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> mesh = TensorMesh([[(1, 6)], [(1, 3)]])
-            >>> fig = plt.figure(figsize=(10, 10))
-            >>> ax1 = fig.add_subplot(211)
-            >>> mesh.plot_grid(ax=ax1)
-            >>> ax1.plot(
-            ...     mesh.cell_centers[:, 0], mesh.cell_centers[:, 1], "ro", markersize=8
-            ... )
-            >>> for ii, loc in zip(range(mesh.nC), mesh.cell_centers):
-            ...     ax1.text(loc[0]+0.05, loc[1]+0.02, "{0:d}".format(ii), color="r")
-            >>> ax1.plot(
-            ...     mesh.faces_x[:, 0], mesh.faces_x[:, 1], "g>", markersize=8
-            ... )
-            >>> for ii, loc in zip(range(mesh.nFx), mesh.faces_x):
-            ...     ax1.text(loc[0]+0.05, loc[1]+0.02, "{0:d}".format(ii), color="g")
-            >>> ax1.plot(
-            ...     mesh.faces_y[:, 0], mesh.faces_y[:, 1], "g^", markersize=8
-            ... )
-            >>> for ii, loc in zip(range(mesh.nFy), mesh.faces_y):
-            ...     ax1.text(loc[0] + 0.05, loc[1] + 0.1, "{0:d}".format((ii + mesh.nFx)), color="g")
-
-            >>> ax1.set_xticks([])
-            >>> ax1.set_yticks([])
-            >>> ax1.spines['bottom'].set_color('white')
-            >>> ax1.spines['top'].set_color('white')
-            >>> ax1.spines['left'].set_color('white')
-            >>> ax1.spines['right'].set_color('white')
-            >>> ax1.set_xlabel('X', fontsize=16, labelpad=-5)
-            >>> ax1.set_ylabel('Y', fontsize=16, labelpad=-15)
-            >>> ax1.set_title("Mapping of Face Divergence", fontsize=14, pad=15)
-            >>> ax1.legend(
-            ...     ['Mesh', '$\\mathbf{\\phi}$ (centers)', '$\\mathbf{u}$ (faces)'],
-            ...     loc='upper right', fontsize=14
-            ... )
-            >>> ax2 = fig.add_subplot(212)
-            >>> ax2.spy(mesh.face_divergence)
-            >>> ax2.set_title("Spy Plot", fontsize=14, pad=5)
-            >>> ax2.set_ylabel("Cell Index", fontsize=12)
-            >>> ax2.set_xlabel("Face Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_face_divergence", None) is None:
             # Get the stencil of +1, -1's
             D = self._face_divergence_stencil
@@ -844,146 +687,6 @@ class DiffOperators(object):
 
     @property
     def nodal_gradient(self):
-        """Nodal gradient operator (nodes to edges)
-
-        This property constructs the 2nd order numerical gradient operator
-        that maps from nodes to edges. The operator is a sparse matrix
-        :math:`\\mathbf{G_n}` that can be applied as a matrix-vector product
-        to a discrete scalar quantity :math:`\\boldsymbol{\\phi}` that
-        lives on the nodes, i.e.::
-
-            grad_phi = Gn @ phi
-
-        Once constructed, the operator is stored permanently as a property of the mesh.
-
-        Returns
-        -------
-        (n_edges, n_nodes) scipy.sparse.csr_matrix
-            The numerical gradient operator from nodes to edges
-
-        Notes
-        -----
-        In continuous space, the gradient operator is defined as:
-
-        .. math::
-            \\vec{u} = \\nabla \\phi = \\frac{\\partial \\phi}{\\partial x}\\hat{x}
-            + \\frac{\\partial \\phi}{\\partial y}\\hat{y}
-            + \\frac{\\partial \\phi}{\\partial z}\\hat{z}
-
-        Where :math:`\\boldsymbol{\\phi}` is the discrete representation of the continuous variable
-        :math:`\\phi` on the nodes and :math:`\\mathbf{u}` is the discrete
-        representation of :math:`\\vec{u}` on the edges, **nodal_gradient** constructs a
-        discrete linear operator :math:`\\mathbf{G_n}` such that:
-
-        .. math::
-            \\mathbf{u} = \\mathbf{G_n} \\, \\boldsymbol{\\phi}
-
-        The Cartesian components of :math:`\\vec{u}` are defined on their corresponding
-        edges (x, y or z) as follows; e.g. the x-component of the gradient is defined
-        on x-edges. For edge :math:`i` which defines a straight path
-        of length :math:`h_i` between adjacent nodes :math:`n_1` and :math:`n_2`:
-
-        .. math::
-            u_i = \\frac{\\phi_{n_2} - \\phi_{n_1}}{h_i}
-
-        Note that :math:`u_i \\in \\mathbf{u}` may correspond to a value on an
-        x, y or z edge. See the example below.
-
-        Examples
-        --------
-        Below, we demonstrate 1) how to apply the nodal gradient operator to
-        a discrete scalar quantity, and 2) the mapping of the nodal gradient
-        operator and its sparsity. Our example is carried out on a 2D mesh
-        but it can be done equivalently for a 3D mesh.
-
-        We start by importing the necessary packages and modules.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> import matplotlib as mpl
-
-        For a discrete scalar quantity defined on the nodes, we take the
-        gradient by constructing the gradient operator and multiplying
-        as a matrix-vector product.
-
-        Create a uniform grid
-
-        >>> h = np.ones(20)
-        >>> mesh = TensorMesh([h, h], "CC")
-
-        Create a discrete scalar on nodes
-
-        >>> nodes = mesh.nodes
-        >>> phi = np.exp(-(nodes[:, 0] ** 2 + nodes[:, 1] ** 2) / 4 ** 2)
-
-        Construct the gradient operator and apply to vector
-
-        >>> Gn = mesh.nodal_gradient
-        >>> grad_phi = Gn @ phi
-
-        Plot the original function and the gradient
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(13, 6))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(phi, v_type="N", ax=ax1)
-            >>> ax1.set_title("Scalar at nodes", fontsize=14)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(
-            ...     grad_phi, ax=ax2, v_type="E", view="vec",
-            ...     stream_opts={"color": "w", "density": 1.0}
-            ... )
-            >>> ax2.set_yticks([])
-            >>> ax2.set_ylabel("")
-            >>> ax2.set_title("Gradient at edges", fontsize=14)
-            >>> plt.show()
-
-        The nodal gradient operator is a sparse matrix that maps
-        from nodes to edges. To demonstrate this, we construct
-        a small 2D mesh. We then show the ordering of the elements in
-        the original discrete quantity :math:`\\boldsymbol{\\phi}` and its
-        discrete gradient as well as a spy plot.
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> mesh = TensorMesh([[(1, 3)], [(1, 6)]])
-            >>> fig = plt.figure(figsize=(12, 10))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_grid(ax=ax1)
-            >>> ax1.set_title("Mapping of Gradient Operator", fontsize=14, pad=15)
-            >>> ax1.plot(mesh.nodes[:, 0], mesh.nodes[:, 1], "ro", markersize=8)
-            >>> for ii, loc in zip(range(mesh.nN), mesh.nodes):
-            >>>     ax1.text(loc[0] + 0.05, loc[1] + 0.02, "{0:d}".format(ii), color="r")
-
-            >>> ax1.plot(mesh.edges_x[:, 0], mesh.edges_x[:, 1], "g>", markersize=8)
-            >>> for ii, loc in zip(range(mesh.nEx), mesh.edges_x):
-            >>>     ax1.text(loc[0] + 0.05, loc[1] + 0.02, "{0:d}".format(ii), color="g")
-
-            >>> ax1.plot(mesh.edges_y[:, 0], mesh.edges_y[:, 1], "g^", markersize=8)
-            >>> for ii, loc in zip(range(mesh.nEy), mesh.edges_y):
-            >>>     ax1.text(loc[0] + 0.05, loc[1] + 0.02, "{0:d}".format((ii + mesh.nEx)), color="g")
-
-            >>> ax1.set_xticks([])
-            >>> ax1.set_yticks([])
-            >>> ax1.spines['bottom'].set_color('white')
-            >>> ax1.spines['top'].set_color('white')
-            >>> ax1.spines['left'].set_color('white')
-            >>> ax1.spines['right'].set_color('white')
-            >>> ax1.set_xlabel('X', fontsize=16, labelpad=-5)
-            >>> ax1.set_ylabel('Y', fontsize=16, labelpad=-15)
-            >>> ax1.legend(
-            >>>     ['Mesh', '$\\mathbf{\\phi}$ (nodes)', '$\\mathbf{u}$ (edges)'],
-            >>>     loc='upper right', fontsize=14
-            >>> )
-            >>> ax2 = fig.add_subplot(122)
-            >>> ax2.spy(mesh.nodal_gradient)
-            >>> ax2.set_title("Spy Plot", fontsize=14, pad=5)
-            >>> ax2.set_ylabel("Edge Index", fontsize=12)
-            >>> ax2.set_xlabel("Node Index", fontsize=12)
-            >>> plt.plot()
-        """
         if getattr(self, "_nodal_gradient", None) is None:
             G = self._nodal_gradient_stencil
             L = self.edge_lengths
@@ -1078,118 +781,6 @@ class DiffOperators(object):
 
     @property
     def nodal_laplacian(self):
-        """Nodal scalar Laplacian operator (nodes to nodes)
-
-        This property constructs the 2nd order scalar Laplacian operator
-        that maps from nodes to nodes. The operator is a sparse matrix
-        :math:`\\mathbf{L_n}` that can be applied as a matrix-vector product to a
-        discrete scalar quantity :math:`\\boldsymbol{\\phi}` that lives on the
-        nodes, i.e.::
-
-            laplace_phi = Ln @ phi
-
-        The operator ``*`` assumes a zero Neumann boundary condition for the discrete
-        scalar quantity. Once constructed, the operator is stored permanently as
-        a property of the mesh.
-
-        Returns
-        -------
-        (n_nodes, n_nodes) scipy.sparse.csr_matrix
-            The numerical Laplacian operator from nodes to nodes
-
-        Notes
-        -----
-        In continuous space, the scalar Laplacian operator is defined as:
-
-        .. math::
-            \\psi = \\nabla^2 \\phi = \\frac{\\partial^2 \\phi}{\\partial x^2}
-            + \\frac{\\partial^2 \\phi}{\\partial y^2}
-            + \\frac{\\partial^2 \\phi}{\\partial z^2}
-
-        Where :math:`\\boldsymbol{\\phi}` is the discrete representation of the continuous variable
-        :math:`\\phi` on the nodes, and :math:`\\boldsymbol{\\psi}` is the discrete representation
-        of its scalar Laplacian on the nodes, **nodal_laplacian** constructs a
-        discrete linear operator :math:`\\mathbf{L_n}` such that:
-
-        .. math::
-            \\boldsymbol{\\psi} = \\mathbf{L_n} \\, \\boldsymbol{\\phi}
-        """
-        # EXAMPLE INACTIVE BECAUSE OPERATOR IS BROKEN
-        #
-        # Examples
-        # --------
-
-        # Below, we demonstrate how to apply the nodal Laplacian operator to
-        # a discrete scalar quantity, the mapping of the nodal Laplacian operator and
-        # its sparsity. Our example is carried out on a 2D mesh but it can
-        # be done equivalently for a 3D mesh.
-
-        # We start by importing the necessary packages and modules.
-
-        # >>> from discretize import TensorMesh
-        # >>> import numpy as np
-        # >>> import matplotlib.pyplot as plt
-        # >>> import matplotlib as mpl
-
-        # For a discrete scalar quantity defined on the nodes, we take the
-        # Laplacian by constructing the operator and multiplying
-        # as a matrix-vector product.
-
-        # >>> # Create a uniform grid
-        # >>> h = np.ones(20)
-        # >>> mesh = TensorMesh([h, h], "CC")
-        # >>>
-        # >>> # Create a discrete scalar on nodes. The scalar MUST
-        # >>> # respect the zero Neumann boundary condition.
-        # >>> nodes = mesh.nodes
-        # >>> phi = np.exp(-(nodes[:, 0] ** 2 + nodes[:, 1] ** 2) / 4 ** 2)
-        # >>>
-        # >>> # Construct the Laplacian operator and apply to vector
-        # >>> Ln = mesh.nodal_laplacian
-        # >>> laplacian_phi = Ln @ phi
-        # >>>
-        # >>> # Plot
-        # >>> fig = plt.figure(figsize=(13, 6))
-        # >>> ax1 = fig.add_subplot(121)
-        # >>> mesh.plot_image(phi, ax=ax1)
-        # >>> ax1.set_title("Scalar at nodes", fontsize=14)
-        # >>> ax2 = fig.add_subplot(122)
-        # >>> mesh.plot_image(laplacian_phi, ax=ax1)
-        # >>> ax2.set_yticks([])
-        # >>> ax2.set_ylabel("")
-        # >>> ax2.set_title("Laplacian at nodes", fontsize=14)
-        # >>> plt.show()
-
-        # The nodal Laplacian operator is a sparse matrix that maps
-        # from nodes to nodes. To demonstrate this, we construct
-        # a small 2D mesh. We then show the ordering of the nodes
-        # and a spy plot illustrating the sparsity of the operator.
-
-        # >>> mesh = TensorMesh([[(1, 4)], [(1, 4)]])
-        # >>> fig = plt.figure(figsize=(12, 6))
-        # >>> ax1 = fig.add_subplot(211)
-        # >>> mesh.plot_grid(ax=ax1)
-        # >>> ax1.set_title("Ordering of the Nodes", fontsize=14, pad=15)
-        # >>> ax1.plot(mesh.nodes[:, 0], mesh.nodes[:, 1], "ro", markersize=8)
-        # >>> for ii, loc in zip(range(mesh.nN), mesh.nodes):
-        # ...     ax1.text(loc[0] + 0.05, loc[1] + 0.02, "{0:d}".format(ii), color="r")
-        # >>> ax1.set_xticks([])
-        # >>> ax1.set_yticks([])
-        # >>> ax1.spines['bottom'].set_color('white')
-        # >>> ax1.spines['top'].set_color('white')
-        # >>> ax1.spines['left'].set_color('white')
-        # >>> ax1.spines['right'].set_color('white')
-        # >>> ax1.set_xlabel('X', fontsize=16, labelpad=-5)
-        # >>> ax1.set_ylabel('Y', fontsize=16, labelpad=-15)
-        # >>> ax1.legend(
-        # ...     ['Mesh', '$\\mathbf{\\phi}$ (nodes)'],
-        # ...     loc='upper right', fontsize=14
-        # ... )
-        # >>> ax2 = fig.add_subplot(212)
-        # >>> ax2.spy(mesh.nodal_laplacian)
-        # >>> ax2.set_title("Spy Plot", fontsize=14, pad=5)
-        # >>> ax2.set_ylabel("Node Index", fontsize=12)
-        # >>> ax2.set_xlabel("Node Index", fontsize=12)
         if getattr(self, "_nodal_laplacian", None) is None:
             warnings.warn("Laplacian has not been tested rigorously.")
             # Compute divergence operator on faces
@@ -1868,136 +1459,6 @@ class DiffOperators(object):
 
     @property
     def stencil_cell_gradient(self):
-        """Stencil for cell gradient operator (cell centers to faces)
-
-        This property constructs a differencing operator that acts on
-        cell centered quantities. The operator takes the difference between
-        the values at adjacent cell centers along each axis direction,
-        and places the result on the shared face; e.g. differences
-        along the x-axis are mapped to x-faces. The operator is a sparse
-        matrix :math:`\\mathbf{G}` that can be applied as a matrix-vector
-        product to a cell centered quantity :math:`\\boldsymbol{\\phi}`, i.e.::
-
-            diff_phi = G @ phi
-
-        By default, the operator assumes zero-Neumann boundary conditions
-        on the scalar quantity. Before calling **stencil_cell_gradient** however,
-        the user can set a mix of zero Dirichlet and zero Neumann boundary
-        conditions using :py:attr:`~discretize.operators.DiffOperators.set_cell_gradient_BC`.
-        When **stencil_cell_gradient** is called, the boundary conditions are
-        enforced for the differencing operator. Once constructed,
-        the operator is stored as a property of the mesh.
-
-        Returns
-        -------
-        (n_faces, n_cells) scipy.sparse.csr_matrix
-            The stencil for the cell gradient
-
-        Examples
-        --------
-        Below, we demonstrate how to set boundary conditions for the cell gradient
-        stencil, construct the cell gradient stencil and apply it to a discrete
-        scalar quantity. The mapping of the cell gradient operator and
-        its sparsity is also illustrated. Our example is carried out on a 2D
-        mesh but it can be done equivalently for a 3D mesh.
-
-        We start by importing the necessary packages and modules.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> import matplotlib as mpl
-
-        We then construct a mesh and define a scalar function at cell
-        centers. In this case, the scalar represents some block within
-        a homogeneous medium.
-
-        Create a uniform grid
-
-        >>> h = np.ones(40)
-        >>> mesh = TensorMesh([h, h], "CC")
-
-        Create a discrete scalar at cell centers
-
-        >>> centers = mesh.cell_centers
-        >>> phi = np.zeros(mesh.nC)
-        >>> k = (np.abs(mesh.cell_centers[:, 0]) < 10.) & (np.abs(mesh.cell_centers[:, 1]) < 10.)
-        >>> phi[k] = 1.
-
-        Before constructing the operator, we must define
-        the boundary conditions; zero Neumann for our example. Once the
-        operator is created, it is applied as a matrix-vector product.
-
-        >>> mesh.set_cell_gradient_BC(['neumann', 'neumann'])
-        >>> G = mesh.stencil_cell_gradient
-        >>> diff_phi = G @ phi
-
-        Now we plot the original scalar, and the differencing taken along the
-        x and y axes.
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(15, 4.5))
-            >>> ax1 = fig.add_subplot(131)
-            >>> mesh.plot_image(phi, ax=ax1)
-            >>> ax1.set_title("Scalar at cell centers", fontsize=14)
-
-            >>> ax2 = fig.add_subplot(132)
-            >>> mesh.plot_image(diff_phi, ax=ax2, v_type="Fx")
-            >>> ax2.set_yticks([])
-            >>> ax2.set_ylabel("")
-            >>> ax2.set_title("Difference (x-axis)", fontsize=14)
-
-            >>> ax3 = fig.add_subplot(133)
-            >>> mesh.plot_image(diff_phi, ax=ax3, v_type="Fy")
-            >>> ax3.set_yticks([])
-            >>> ax3.set_ylabel("")
-            >>> ax3.set_title("Difference (y-axis)", fontsize=14)
-            >>> plt.show()
-
-        The cell gradient stencil is a sparse differencing matrix that maps
-        from cell centers to faces. To demonstrate this, we construct
-        a small 2D mesh. We then show the ordering of the elements
-        and a spy plot.
-
-        >>> mesh = TensorMesh([[(1, 3)], [(1, 6)]])
-        >>> mesh.set_cell_gradient_BC('neumann')
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(12, 10))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_grid(ax=ax1)
-            >>> ax1.set_title("Mapping of Stencil", fontsize=14, pad=15)
-            >>> ax1.plot(mesh.cell_centers[:, 0], mesh.cell_centers[:, 1], "ro", markersize=8)
-            >>> for ii, loc in zip(range(mesh.nC), mesh.cell_centers):
-            ...     ax1.text(loc[0] + 0.05, loc[1] + 0.02, "{0:d}".format(ii), color="r")
-            >>> ax1.plot(mesh.faces_x[:, 0], mesh.faces_x[:, 1], "g>", markersize=8)
-            >>> for ii, loc in zip(range(mesh.nFx), mesh.faces_x):
-            ...     ax1.text(loc[0] + 0.05, loc[1] + 0.02, "{0:d}".format(ii), color="g")
-            >>> ax1.plot(mesh.faces_y[:, 0], mesh.faces_y[:, 1], "g^", markersize=8)
-            >>> for ii, loc in zip(range(mesh.nFy), mesh.faces_y):
-            ...     ax1.text(loc[0] + 0.05, loc[1] + 0.02, "{0:d}".format((ii + mesh.nFx)), color="g")
-            >>> ax1.set_xticks([])
-            >>> ax1.set_yticks([])
-            >>> ax1.spines['bottom'].set_color('white')
-            >>> ax1.spines['top'].set_color('white')
-            >>> ax1.spines['left'].set_color('white')
-            >>> ax1.spines['right'].set_color('white')
-            >>> ax1.set_xlabel('X', fontsize=16, labelpad=-5)
-            >>> ax1.set_ylabel('Y', fontsize=16, labelpad=-15)
-            >>> ax1.legend(
-            ...     ['Mesh', r'$\\mathbf{\\phi}$ (centers)', r'$\\mathbf{G^\\ast \\phi}$ (faces)'],
-            ...     loc='upper right', fontsize=14
-            ... )
-
-            >>> ax2 = fig.add_subplot(122)
-            >>> ax2.spy(mesh.stencil_cell_gradient)
-            >>> ax2.set_title("Spy Plot", fontsize=14, pad=5)
-            >>> ax2.set_ylabel("Face Index", fontsize=12)
-            >>> ax2.set_xlabel("Cell Index", fontsize=12)
-            >>> plt.show()
-        """
         BC = self.set_cell_gradient_BC(self._cell_gradient_BC_list)
         if self.dim == 1:
             G = _ddxCellGrad(self.shape_cells[0], BC[0])
@@ -2794,158 +2255,6 @@ class DiffOperators(object):
 
     @property
     def edge_curl(self):
-        r"""Edge curl operator (edges to faces)
-
-        This property constructs the 2nd order numerical curl operator
-        that maps from edges to faces. The operator is a sparse matrix
-        :math:`\mathbf{C_e}` that can be applied as a matrix-vector product
-        to a discrete vector quantity **u** that lives
-        on the edges; i.e.::
-
-            curl_u = Ce @ u
-
-        Once constructed, the operator is stored permanently as a property of the mesh.
-
-        Returns
-        -------
-        (n_faces, n_edges) scipy.sparse.csr_matrix
-            The numerical curl operator from edges to faces
-
-        Notes
-        -----
-        In continuous space, the curl operator is defined as:
-
-        .. math::
-            \vec{w} = \nabla \times \vec{u} =
-            \begin{vmatrix}
-            \hat{x} & \hat{y} & \hat{z} \\
-            \partial_x & \partial_y & \partial_z \\
-            u_x & u_y & u_z
-            \end{vmatrix}
-
-        Where :math:`\mathbf{u}` is the discrete representation of the continuous variable
-        :math:`\vec{u}` on cell edges and :math:`\mathbf{w}` is the discrete
-        representation of the curl on the faces, **edge_curl** constructs a
-        discrete linear operator :math:`\\mathbf{C_e}` such that:
-
-        .. math::
-            \mathbf{w} = \mathbf{C_e \, u}
-
-        The computation of the curl on mesh faces can be expressed
-        according to the integral form below. For face :math:`i` bordered by
-        a set of edges indexed by subset :math:`K`:
-
-        .. math::
-            w_i = \frac{1}{A_i} \sum_{k \in K} \vec{u}_k \cdot \vec{\ell}_k
-
-        where :math:`A_i` is the surface area of face *i*,
-        :math:`u_k` is the value of :math:`\vec{u}` on face *k*,
-        and \vec{\ell}_k is the path along edge *k*.
-
-        Examples
-        --------
-        Below, we demonstrate the mapping and sparsity of the edge curl
-        for a 3D tensor mesh. We choose a the index for a single face,
-        and illustrate which edges are used to compute the curl on that
-        face.
-
-        >>> from discretize import TensorMesh
-        >>> from discretize.utils import mkvc
-        >>> import matplotlib.pyplot as plt
-        >>> import numpy as np
-        >>> import matplotlib as mpl
-        >>> import mpl_toolkits.mplot3d as mp3d
-        >>> mpl.rcParams.update({'font.size': 14})
-
-        Create a simple tensor mesh, and grab the **edge_curl** operator:
-
-        >>> mesh = TensorMesh([[(1, 2)], [(1, 2)], [(1, 2)]])
-        >>> Ce = mesh.edge_curl
-
-        Then we choose a *face* for illustration purposes:
-
-        >>> face_ind = 2  # Index of a face in the mesh (could be x, y or z)
-        >>> edge_ind = np.where(
-        ...     np.sum((mesh.edges-mesh.faces[face_ind, :])**2, axis=1) <= 0.5 + 1e-6
-        ... )[0]
-
-        >>> face = mesh.faces[face_ind, :]
-        >>> face_norm = mesh.face_normals[face_ind, :]
-        >>> edges = mesh.edges[edge_ind, :]
-        >>> edge_tan = mesh.edge_tangents[edge_ind, :]
-        >>> node = np.min(edges, axis=0)
-
-        >>> min_edges = np.min(edges, axis=0)
-        >>> max_edges = np.max(edges, axis=0)
-        >>> if face_norm[0] == 1:
-        ...     k = (edges[:, 1] == min_edges[1]) | (edges[:, 2] == max_edges[2])
-        ...     poly = node + np.c_[np.r_[0, 0, 0, 0], np.r_[0, 1, 1, 0], np.r_[0, 0, 1, 1]]
-        ...     ds = [0.07, -0.07, -0.07]
-        ... elif face_norm[1] == 1:
-        ...     k = (edges[:, 0] == max_edges[0]) | (edges[:, 2] == min_edges[2])
-        ...     poly = node + np.c_[np.r_[0, 1, 1, 0], np.r_[0, 0, 0, 0], np.r_[0, 0, 1, 1]]
-        ...     ds = [0.07, -0.09, -0.07]
-        ... elif face_norm[2] == 1:
-        ...     k = (edges[:, 0] == min_edges[0]) | (edges[:, 1] == max_edges[1])
-        ...     poly = node + np.c_[np.r_[0, 1, 1, 0], np.r_[0, 0, 1, 1], np.r_[0, 0, 0, 0]]
-        ...     ds = [0.07, -0.09, -0.07]
-        >>> edge_tan[k, :] *= -1
-
-        Plot the curve and its mapping for a single face.
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(10, 15))
-            >>> ax1 = fig.add_axes([0, 0.35, 1, 0.6], projection='3d', elev=25, azim=-60)
-            >>> mesh.plot_grid(ax=ax1)
-            >>> ax1.plot(
-            ...     mesh.edges[edge_ind, 0], mesh.edges[edge_ind, 1], mesh.edges[edge_ind, 2],
-            ...     "go", markersize=10
-            ... )
-            >>> ax1.plot(
-            ...     mesh.faces[face_ind, 0], mesh.faces[face_ind, 1], mesh.faces[face_ind, 2],
-            ...     "rs", markersize=10
-            ... )
-            >>> poly = mp3d.art3d.Poly3DCollection(
-            ...     [poly], alpha=0.1, facecolor='r', linewidth=None
-            ... )
-            >>> ax1.add_collection(poly)
-            >>> ax1.quiver(
-            ...     edges[:, 0], edges[:, 1], edges[:, 2],
-            ...     0.5*edge_tan[:, 0], 0.5*edge_tan[:, 1], 0.5*edge_tan[:, 2],
-            ...     edgecolor='g', pivot='middle', linewidth=4, arrow_length_ratio=0.25
-            ... )
-            >>> ax1.text(face[0]+ds[0], face[1]+ds[1], face[2]+ds[2], "{0:d}".format(face_ind), color="r")
-            >>> for ii, loc in zip(range(len(edge_ind)), edges):
-            ...     ax1.text(loc[0]+ds[0], loc[1]+ds[1], loc[2]+ds[2], "{0:d}".format(edge_ind[ii]), color="g")
-            >>> ax1.legend(
-            ...     ['Mesh', '$\\mathbf{u}$ (edges)', '$\\mathbf{w}$ (face)'],
-            ...     loc='upper right', fontsize=14
-            ... )
-
-            Manually make axis properties invisible
-
-            >>> ax1.set_xticks([])
-            >>> ax1.set_yticks([])
-            >>> ax1.set_zticks([])
-            >>> ax1.w_xaxis.set_pane_color((1.0, 1.0, 1.0, 0.0))
-            >>> ax1.w_yaxis.set_pane_color((1.0, 1.0, 1.0, 0.0))
-            >>> ax1.w_zaxis.set_pane_color((1.0, 1.0, 1.0, 0.0))
-            >>> ax1.w_xaxis.line.set_color((1.0, 1.0, 1.0, 0.0))
-            >>> ax1.w_yaxis.line.set_color((1.0, 1.0, 1.0, 0.0))
-            >>> ax1.w_zaxis.line.set_color((1.0, 1.0, 1.0, 0.0))
-            >>> ax1.set_xlabel('X', labelpad=-15, fontsize=16)
-            >>> ax1.set_ylabel('Y', labelpad=-20, fontsize=16)
-            >>> ax1.set_zlabel('Z', labelpad=-20, fontsize=16)
-            >>> ax1.set_title("Mapping for a Single Face", fontsize=16, pad=-15)
-
-            >>> ax2 = fig.add_axes([0.05, 0.05, 0.9, 0.3])
-            >>> ax2.spy(Ce)
-            >>> ax2.set_title("Spy Plot", fontsize=16, pad=5)
-            >>> ax2.set_ylabel("Face Index", fontsize=12)
-            >>> ax2.set_xlabel("Edge Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_edge_curl", None) is None:
             if self.dim <= 1:
                 raise NotImplementedError("Edge Curl only programed for 2 or 3D.")
@@ -2959,27 +2268,6 @@ class DiffOperators(object):
 
     @property
     def boundary_face_scalar_integral(self):
-        r"""Represents the operation of integrating a scalar function on the boundary
-
-        This matrix represents the boundary surface integral of a scalar function
-        multiplied with a finite volume test function on the mesh.
-
-        Returns
-        -------
-        (n_faces, n_boundary_faces) scipy.sparse.csr_matrix
-
-        Notes
-        -----
-        The integral we are representing on the boundary of the mesh is
-
-        .. math:: \int_{\Omega} u\vec{w} \cdot \hat{n} \partial \Omega
-
-        In discrete form this is:
-
-        .. math:: w^T * P @ u_b
-
-        where `w` is defined on all faces, and `u_b` is defined on boundary faces.
-        """
         if self.dim == 1:
             return sp.csr_matrix(
                 ([-1, 1], ([0, self.n_faces_x - 1], [0, 1])), shape=(self.n_faces_x, 2)
@@ -3272,90 +2560,6 @@ class DiffOperators(object):
 
     @property
     def average_face_to_cell(self):
-        """Averaging operator from faces to cell centers (scalar quantities).
-
-        This property constructs a 2nd order averaging operator that maps scalar
-        quantities from faces to cell centers. This averaging operator is
-        used when a discrete scalar quantity defined on mesh faces must be
-        projected to cell centers. Once constructed, the operator is stored
-        permanently as a property of the mesh. *See notes*.
-
-        Returns
-        -------
-        (n_cells, n_faces) scipy.sparse.csr_matrix
-            The scalar averaging operator from faces to cell centers
-
-        Notes
-        -----
-        Let :math:`\\boldsymbol{\\phi_f}` be a discrete scalar quantity that
-        lives on mesh faces. **average_face_to_cell** constructs a discrete
-        linear operator :math:`\\mathbf{A_{fc}}` that projects
-        :math:`\\boldsymbol{\\phi_f}` to cell centers, i.e.:
-
-        .. math::
-            \\boldsymbol{\\phi_c} = \\mathbf{A_{fc}} \\, \\boldsymbol{\\phi_f}
-
-        where :math:`\\boldsymbol{\\phi_c}` approximates the value of the scalar
-        quantity at cell centers. For each cell, we are simply averaging
-        the values defined on its faces. The operation is implemented as a
-        matrix vector product, i.e.::
-
-            phi_c = Afc @ phi_f
-
-        Examples
-        --------
-        Here we compute the values of a scalar function on the faces. We then create
-        an averaging operator to approximate the function at cell centers. We choose
-        to define a scalar function that is strongly discontinuous in some places to
-        demonstrate how the averaging operator will smooth out discontinuities.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-
-        >>> h = np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Then we create a scalar variable on faces
-
-        >>> phi_f = np.zeros(mesh.nF)
-        >>> xy = mesh.faces
-        >>> phi_f[(xy[:, 1] > 0)] = 25.0
-        >>> phi_f[(xy[:, 1] < -10.0) & (xy[:, 0] > -10.0) & (xy[:, 0] < 10.0)] = 50.0
-
-        Next, we construct the averaging operator and apply it to
-        the discrete scalar quantity to approximate the value at cell centers.
-
-        >>> Afc = mesh.average_face_to_cell
-        >>> phi_c = Afc @ phi_f
-
-        And finally plot the results:
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(phi_f, ax=ax1, v_type="F")
-            >>> ax1.set_title("Variable at faces", fontsize=16)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(phi_c, ax=ax2, v_type="CC")
-            >>> ax2.set_title("Averaged to cell centers", fontsize=16)
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Afc, ms=1)
-            >>> ax1.set_title("Face Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Cell Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_face_to_cell", None) is None:
             if self.dim == 1:
                 self._average_face_to_cell = self.aveFx2CC
@@ -3371,105 +2575,6 @@ class DiffOperators(object):
 
     @property
     def average_face_to_cell_vector(self):
-        """Averaging operator from faces to cell centers (vector quantities).
-
-        This property constructs the averaging operator that independently maps the
-        Cartesian components of vector quantities from faces to cell centers.
-        This averaging operators is used when a discrete vector quantity defined on mesh faces
-        must be approximated at cell centers. Once constructed, the operator is
-        stored permanently as a property of the mesh.
-
-        Be aware that the Cartesian components of the original vector
-        are defined on their respective faces; e.g. the x-component lives
-        on x-faces. However, the x, y and z components are being averaged
-        separately to cell centers. The operation is implemented as a
-        matrix vector product, i.e.::
-
-            u_c = Afc @ u_f
-
-        Returns
-        -------
-        (dim * n_cells, n_faces) scipy.sparse.csr_matrix
-            The vector averaging operator from faces to cell centers. Since we
-            are averaging a vector quantity to cell centers, the first dimension
-            of the operator is the mesh dimension times the number of cells.
-
-        Notes
-        -----
-        Let :math:`\\mathbf{u_f}` be the discrete representation of a vector
-        quantity whose Cartesian components are defined on their respective faces;
-        e.g. :math:`u_x` is defined on x-faces.
-        **average_face_to_cell_vector** constructs a discrete linear operator
-        :math:`\\mathbf{A_{fc}}` that projects each Cartesian component of
-        :math:`\\mathbf{u_f}` independently to cell centers, i.e.:
-
-        .. math::
-            \\mathbf{u_c} = \\mathbf{A_{fc}} \\, \\mathbf{u_f}
-
-        where :math:`\\mathbf{u_c}` is a discrete vector quantity whose Cartesian
-        components defined at the cell centers and organized into a 1D array of
-        the form np.r_[ux, uy, uz]. For each cell, and for each Cartesian component,
-        we are simply taking the average of the values
-        defined on the cell's corresponding faces and placing the result at
-        the cell's center.
-
-        Examples
-        --------
-        Here we compute the values of a vector function discretized to the mesh faces.
-        We then create an averaging operator to approximate the function at cell centers.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-
-        >>> h = 0.5 * np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Then we create a discrete vector on mesh faces
-
-        >>> faces_x = mesh.faces_x
-        >>> faces_y = mesh.faces_y
-        >>> u_fx = -(faces_x[:, 1] / np.sqrt(np.sum(faces_x ** 2, axis=1))) * np.exp(
-        >>>     -(faces_x[:, 0] ** 2 + faces_x[:, 1] ** 2) / 6 ** 2
-        >>> )
-        >>> u_fy = (faces_y[:, 0] / np.sqrt(np.sum(faces_y ** 2, axis=1))) * np.exp(
-        >>>     -(faces_y[:, 0] ** 2 + faces_y[:, 1] ** 2) / 6 ** 2
-        >>> )
-        >>> u_f = np.r_[u_fx, u_fy]
-
-        Next, we construct the averaging operator and apply it to
-        the discrete vector quantity to approximate the value at cell centers.
-
-        >>> Afc = mesh.average_face_to_cell_vector
-        >>> u_c = Afc @ u_f
-
-        And finally, plot the results:
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(u_f, ax=ax1, v_type="F", view='vec')
-            >>> ax1.set_title("Variable at faces", fontsize=16)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(u_c, ax=ax2, v_type="CCv", view='vec')
-            >>> ax2.set_title("Averaged to cell centers", fontsize=16)
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Afc, ms=1)
-            >>> ax1.set_title("Face Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Cell Vector Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_face_to_cell_vector", None) is None:
             if self.dim == 1:
                 self._average_face_to_cell_vector = self.aveFx2CC
@@ -3777,98 +2882,6 @@ class DiffOperators(object):
 
     @property
     def average_cell_to_face(self):
-        """Averaging operator from cell centers to faces (scalar quantities).
-
-        This property constructs an averaging operator that maps scalar
-        quantities from cell centers to face. This averaging operator is
-        used when a discrete scalar quantity defined cell centers must be
-        projected to faces. Once constructed, the operator is stored
-        permanently as a property of the mesh. *See notes*.
-
-        Returns
-        -------
-        (n_faces, n_cells) scipy.sparse.csr_matrix
-            The scalar averaging operator from cell centers to faces
-
-        Notes
-        -----
-        Let :math:`\\boldsymbol{\\phi_c}` be a discrete scalar quantity that
-        lives at cell centers. **average_cell_to_face** constructs a discrete
-        linear operator :math:`\\mathbf{A_{cf}}` that projects
-        :math:`\\boldsymbol{\\phi_c}` to faces, i.e.:
-
-        .. math::
-            \\boldsymbol{\\phi_f} = \\mathbf{A_{cf}} \\, \\boldsymbol{\\phi_c}
-
-        where :math:`\\boldsymbol{\\phi_f}` approximates the value of the scalar
-        quantity at the faces. For each face, we are performing a weighted average
-        between the values at adjacent cell centers. In 1D, where adjacent cells
-        :math:`i` and :math:`i+1` have widths :math:`h_i` and :math:`h_{i+1}`,
-        :math:`\\phi` on face is approximated by:
-
-        .. math::
-            \\phi_{i \\! + \\! 1/2} \\approx \\frac{h_{i+1} \\phi_i + h_i \\phi_{i+1}}{h_i + h_{i+1}}
-
-        On boundary faces, nearest neighbour is used to extrapolate the value
-        from the nearest cell center. Once the operator is construct, the averaging
-        is implemented as a matrix vector product, i.e.::
-
-            phi_f = Acf @ phi_c
-
-        Examples
-        --------
-        Here we compute the values of a scalar function at cell centers. We then create
-        an averaging operator to approximate the function on the faces. We choose
-        to define a scalar function that is strongly discontinuous in some places to
-        demonstrate how the averaging operator will smooth out discontinuities.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-
-        >>> h = np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Create a scalar variable at cell centers
-
-        >>> phi_c = np.zeros(mesh.nC)
-        >>> xy = mesh.cell_centers
-        >>> phi_c[(xy[:, 1] > 0)] = 25.0
-        >>> phi_c[(xy[:, 1] < -10.0) & (xy[:, 0] > -10.0) & (xy[:, 0] < 10.0)] = 50.0
-
-        Next, we construct the averaging operator and apply it to
-        the discrete scalar quantity to approximate the value at the faces.
-
-        >>> Acf = mesh.average_cell_to_face
-        >>> phi_f = Acf @ phi_c
-
-        Plot the results
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(phi_c, ax=ax1, v_type="CC")
-            >>> ax1.set_title("Variable at cell centers", fontsize=16)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(phi_f, ax=ax2, v_type="F")
-            >>> ax2.set_title("Averaged to faces", fontsize=16)
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator.
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Acf, ms=1)
-            >>> ax1.set_title("Cell Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Face Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_cell_to_face", None) is None:
             if self.dim == 1:
                 self._average_cell_to_face = av_extrap(self.shape_cells[0])
@@ -3909,110 +2922,6 @@ class DiffOperators(object):
 
     @property
     def average_cell_vector_to_face(self):
-        """Averaging operator from cell centers to faces (vector quantities).
-
-        This property constructs the averaging operator that independently maps the
-        Cartesian components of vector quantities from cell centers to faces.
-        This averaging operators is used when a discrete vector quantity defined at
-        cell centers must be approximated on the faces. Once constructed, the operator is
-        stored permanently as a property of the mesh.
-
-        Be aware that the Cartesian components of the original vector
-        are defined seperately at cell centers in a 1D numpy.array organized [ux, uy, uz].
-        Once projected to faces, the Cartesian components are defined on their respective
-        faces; e.g. the x-component lives on x-faces. The operation is implemented as a
-        matrix vector product, i.e.::
-
-            u_f = Acf @ u_c
-
-        Returns
-        -------
-        (n_faces, dim * n_cells) scipy.sparse.csr_matrix
-            The vector averaging operator from cell centers to faces. Since we
-            are averaging a vector quantity from cell centers, the second dimension
-            of the operator is the mesh dimension times the number of cells.
-
-        Notes
-        -----
-        Let :math:`\\mathbf{u_c}` be the discrete representation of a vector
-        quantity whose Cartesian components are defined separately at cell centers.
-        **average_cell_vector_to_face** constructs a discrete linear operator
-        :math:`\\mathbf{A_{cf}}` that projects each Cartesian component of
-        :math:`\\mathbf{u_c}` to the faces, i.e.:
-
-        .. math::
-            \\mathbf{u_f} = \\mathbf{A_{cf}} \\, \\mathbf{u_c}
-
-        where :math:`\\mathbf{u_f}` is the discrete vector quantity whose Cartesian
-        components are approximated on their respective cell faces; e.g. the x-component is
-        approximated on x-faces. For each face (x, y or z), we are simply taking a weighted average
-        between the values of the correct Cartesian component at the corresponding cell centers.
-
-        E.g. for the x-component, which is projected to x-faces, the weighted average on
-        a 2D mesh would be:
-
-        .. math::
-            u_x(i \\! + \\! 1/2, j) = \\frac{h_{i+1} u_x (i,j) + h_i u_x(i \\! + \\! 1,j)}{hx_i + hx_{i+1}}
-
-        where :math:`h_i` and :math:`h_{i+1}` represent the cell respective cell widths
-        in the x-direction. For boundary faces, nearest neighbor is used to extrapolate
-        the values.
-
-        Examples
-        --------
-        Here we compute the values of a vector function discretized to cell centers.
-        We then create an averaging operator to approximate the function on the faces.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> h = 0.5 * np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Then we create a discrete vector at cell centers,
-
-        >>> centers = mesh.cell_centers
-        >>> u_x = -(centers[:, 1] / np.sqrt(np.sum(centers ** 2, axis=1))) * np.exp(
-        ...     -(centers[:, 0] ** 2 + centers[:, 1] ** 2) / 6 ** 2
-        ... )
-        >>> u_y = (centers[:, 0] / np.sqrt(np.sum(centers ** 2, axis=1))) * np.exp(
-        ...     -(centers[:, 0] ** 2 + centers[:, 1] ** 2) / 6 ** 2
-        ... )
-        >>> u_c = np.r_[u_x, u_y]
-
-        Next, we construct the averaging operator and apply it to
-        the discrete vector quantity to approximate the value on the faces.
-
-        >>> Acf = mesh.average_cell_vector_to_face
-        >>> u_f = Acf @ u_c
-
-        And plot the results
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(u_c, ax=ax1, v_type="CCv", view='vec')
-            >>> ax1.set_title("Variable at faces", fontsize=16)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(u_f, ax=ax2, v_type="F", view='vec')
-            >>> ax2.set_title("Averaged to cell centers", fontsize=16)
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Acf, ms=1)
-            >>> ax1.set_title("Cell Vector Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Face Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_cell_vector_to_face", None) is None:
             if self.dim == 1:
                 self._average_cell_vector_to_face = self.aveCC2F
@@ -4049,97 +2958,6 @@ class DiffOperators(object):
 
     @property
     def average_cell_to_edge(self):
-        """Averaging operator from cell centers to edges (scalar quantities).
-
-        This property constructs an averaging operator that maps scalar
-        quantities from cell centers to edge. This averaging operator is
-        used when a discrete scalar quantity defined cell centers must be
-        projected to edges. Once constructed, the operator is stored
-        permanently as a property of the mesh. *See notes*.
-
-        Returns
-        -------
-        (n_edges, n_cells) scipy.sparse.csr_matrix
-            The scalar averaging operator from edges to cell centers
-
-        Notes
-        -----
-        Let :math:`\\boldsymbol{\\phi_c}` be a discrete scalar quantity that
-        lives at cell centers. **average_cell_to_edge** constructs a discrete
-        linear operator :math:`\\mathbf{A_{ce}}` that projects
-        :math:`\\boldsymbol{\\phi_c}` to edges, i.e.:
-
-        .. math::
-            \\boldsymbol{\\phi_e} = \\mathbf{A_{ce}} \\, \\boldsymbol{\\phi_c}
-
-        where :math:`\\boldsymbol{\\phi_e}` approximates the value of the scalar
-        quantity at the edges. For each edge, we are performing a weighted average
-        between the values at adjacent cell centers. In 1D, where adjacent cells
-        :math:`i` and :math:`i+1` have widths :math:`h_i` and :math:`h_{i+1}`,
-        :math:`\\phi` on edge (node location in 1D) is approximated by:
-
-        .. math::
-            \\phi_{i \\! + \\! 1/2} \\approx \\frac{h_{i+1} \\phi_i + h_i \\phi_{i+1}}{h_i + h_{i+1}}
-
-        On boundary edges, nearest neighbour is used to extrapolate the value
-        from the nearest cell center. Once the operator is construct, the averaging
-        is implemented as a matrix vector product, i.e.::
-
-            phi_e = Ace @ phi_c
-
-        Examples
-        --------
-        Here we compute the values of a scalar function at cell centers. We then create
-        an averaging operator to approximate the function on the edges. We choose
-        to define a scalar function that is strongly discontinuous in some places to
-        demonstrate how the averaging operator will smooth out discontinuities.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> h = np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Then we create a scalar variable at cell centers
-
-        >>> phi_c = np.zeros(mesh.nC)
-        >>> xy = mesh.cell_centers
-        >>> phi_c[(xy[:, 1] > 0)] = 25.0
-        >>> phi_c[(xy[:, 1] < -10.0) & (xy[:, 0] > -10.0) & (xy[:, 0] < 10.0)] = 50.0
-
-        Next, we construct the averaging operator and apply it to
-        the discrete scalar quantity to approximate the value at the edges.
-
-        >>> Ace = mesh.average_cell_to_edge
-        >>> phi_e = Ace @ phi_c
-
-        And plot the results:
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(phi_c, ax=ax1, v_type="CC")
-            >>> ax1.set_title("Variable at cell centers", fontsize=16)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(phi_e, ax=ax2, v_type="E")
-            >>> ax2.set_title("Averaged to edges", fontsize=16)
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator.
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Ace, ms=1)
-            >>> ax1.set_title("Cell Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Edge Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_cell_to_edge", None) is None:
             n = self.shape_cells
             if self.dim == 1:
@@ -4166,89 +2984,6 @@ class DiffOperators(object):
 
     @property
     def average_edge_to_cell(self):
-        """Averaging operator from edges to cell centers (scalar quantities).
-
-        This property constructs a 2nd order averaging operator that maps scalar
-        quantities from edges to cell centers. This averaging operator is
-        used when a discrete scalar quantity defined on mesh edges must be
-        projected to cell centers. Once constructed, the operator is stored
-        permanently as a property of the mesh. *See notes*.
-
-        Returns
-        -------
-        (n_cells, n_edges) scipy.sparse.csr_matrix
-            The scalar averaging operator from edges to cell centers
-
-        Notes
-        -----
-        Let :math:`\\boldsymbol{\\phi_e}` be a discrete scalar quantity that
-        lives on mesh edges. **average_edge_to_cell** constructs a discrete
-        linear operator :math:`\\mathbf{A_{ec}}` that projects
-        :math:`\\boldsymbol{\\phi_e}` to cell centers, i.e.:
-
-        .. math::
-            \\boldsymbol{\\phi_c} = \\mathbf{A_{ec}} \\, \\boldsymbol{\\phi_e}
-
-        where :math:`\\boldsymbol{\\phi_c}` approximates the value of the scalar
-        quantity at cell centers. For each cell, we are simply averaging
-        the values defined on its edges. The operation is implemented as a
-        matrix vector product, i.e.::
-
-            phi_c = Aec @ phi_e
-
-        Examples
-        --------
-        Here we compute the values of a scalar function on the edges. We then create
-        an averaging operator to approximate the function at cell centers. We choose
-        to define a scalar function that is strongly discontinuous in some places to
-        demonstrate how the averaging operator will smooth out discontinuities.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> h = np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Then we create a scalar variable on edges,
-
-        >>> phi_e = np.zeros(mesh.nE)
-        >>> xy = mesh.edges
-        >>> phi_e[(xy[:, 1] > 0)] = 25.0
-        >>> phi_e[(xy[:, 1] < -10.0) & (xy[:, 0] > -10.0) & (xy[:, 0] < 10.0)] = 50.0
-
-        Next, we construct the averaging operator and apply it to
-        the discrete scalar quantity to approximate the value at cell centers.
-
-        >>> Aec = mesh.average_edge_to_cell
-        >>> phi_c = Aec @ phi_e
-
-        And plot the results:
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(phi_e, ax=ax1, v_type="E")
-            >>> ax1.set_title("Variable at edges", fontsize=16)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(phi_c, ax=ax2, v_type="CC")
-            >>> ax2.set_title("Averaged to cell centers", fontsize=16)
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Aec, ms=1)
-            >>> ax1.set_title("Edge Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Cell Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_edge_to_cell", None) is None:
             if self.dim == 1:
                 self._avE2CC = self.aveEx2CC
@@ -4264,104 +2999,6 @@ class DiffOperators(object):
 
     @property
     def average_edge_to_cell_vector(self):
-        """Averaging operator from edges to cell centers (vector quantities).
-
-        This property constructs the averaging operator that independently maps the
-        Cartesian components of vector quantities from edges to cell centers.
-        This averaging operators is used when a discrete vector quantity defined on mesh edges
-        must be approximated at cell centers. Once constructed, the operator is
-        stored permanently as a property of the mesh.
-
-        Be aware that the Cartesian components of the original vector
-        are defined on their respective edges; e.g. the x-component lives
-        on x-edges. However, the x, y and z components are being averaged
-        separately to cell centers. The operation is implemented as a
-        matrix vector product, i.e.::
-
-            u_c = Aec @ u_e
-
-        Returns
-        -------
-        (dim * n_cells, n_edges) scipy.sparse.csr_matrix
-            The vector averaging operator from edges to cell centers. Since we
-            are averaging a vector quantity to cell centers, the first dimension
-            of the operator is the mesh dimension times the number of cells.
-
-        Notes
-        -----
-        Let :math:`\\mathbf{u_e}` be the discrete representation of a vector
-        quantity whose Cartesian components are defined on their respective edges;
-        e.g. :math:`u_x` is defined on x-edges.
-        **average_edge_to_cell_vector** constructs a discrete linear operator
-        :math:`\\mathbf{A_{ec}}` that projects each Cartesian component of
-        :math:`\\mathbf{u_e}` independently to cell centers, i.e.:
-
-        .. math::
-            \\mathbf{u_c} = \\mathbf{A_{ec}} \\, \\mathbf{u_e}
-
-        where :math:`\\mathbf{u_c}` is a discrete vector quantity whose Cartesian
-        components defined at the cell centers and organized into a 1D array of
-        the form np.r_[ux, uy, uz]. For each cell, and for each Cartesian component,
-        we are simply taking the average of the values
-        defined on the cell's corresponding edges and placing the result at
-        the cell's center.
-
-        Examples
-        --------
-        Here we compute the values of a vector function discretized to the edges.
-        We then create an averaging operator to approximate the function at cell centers.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> h = 0.5 * np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Then we create a discrete vector on mesh edges
-
-        >>> edges_x = mesh.edges_x
-        >>> edges_y = mesh.edges_y
-        >>> u_ex = -(edges_x[:, 1] / np.sqrt(np.sum(edges_x ** 2, axis=1))) * np.exp(
-        ...     -(edges_x[:, 0] ** 2 + edges_x[:, 1] ** 2) / 6 ** 2
-        ... )
-        >>> u_ey = (edges_y[:, 0] / np.sqrt(np.sum(edges_y ** 2, axis=1))) * np.exp(
-        ...     -(edges_y[:, 0] ** 2 + edges_y[:, 1] ** 2) / 6 ** 2
-        ... )
-        >>> u_e = np.r_[u_ex, u_ey]
-
-        Next, we construct the averaging operator and apply it to
-        the discrete vector quantity to approximate the value at cell centers.
-
-        >>> Aec = mesh.average_edge_to_cell_vector
-        >>> u_c = Aec @ u_e
-
-        And plot the results:
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(u_e, ax=ax1, v_type="E", view='vec')
-            >>> ax1.set_title("Variable at edges", fontsize=16)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(u_c, ax=ax2, v_type="CCv", view='vec')
-            >>> ax2.set_title("Averaged to cell centers", fontsize=16)
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Aec, ms=1)
-            >>> ax1.set_title("Edge Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Cell Vector Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_edge_to_cell_vector", None) is None:
             if self.dim == 1:
                 self._average_edge_to_cell_vector = self.aveEx2CC
@@ -4668,97 +3305,6 @@ class DiffOperators(object):
 
     @property
     def average_edge_to_face_vector(self):
-        """Averaging operator from edges to faces (vector quantities).
-
-        This property constructs the averaging operator that independently maps the
-        Cartesian components of vector quantities from edges to faces.
-        This averaging operators is used when a discrete vector quantity defined on mesh edges
-        must be approximated at faces. The operation is implemented as a
-        matrix vector product, i.e.::
-
-            u_f = Aef @ u_e
-
-        Once constructed, the operator is stored permanently as a property of the mesh.
-
-        Returns
-        -------
-        (n_faces, n_edges) scipy.sparse.csr_matrix
-            The vector averaging operator from edges to faces.
-
-        Notes
-        -----
-        Let :math:`\\mathbf{u_e}` be the discrete representation of a vector
-        quantity whose Cartesian components are defined on their respective edges;
-        e.g. the x-component is defined on x-edges. **average_edge_to_face_vector**
-        constructs a discrete linear operator :math:`\\mathbf{A_{ef}}` that
-        projects each Cartesian component of :math:`\\mathbf{u_e}` to
-        its corresponding face, i.e.:
-
-        .. math::
-            \\mathbf{u_f} = \\mathbf{A_{ef}} \\, \\mathbf{u_e}
-
-        where :math:`\\mathbf{u_f}` is a discrete vector quantity whose Cartesian
-        components are defined on their respective faces; e.g. the x-component is
-        defined on x-faces.
-
-        Examples
-        --------
-        Here we compute the values of a vector function discretized to the edges.
-        We then create an averaging operator to approximate the function on
-        the faces.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> h = 0.5 * np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Create a discrete vector on mesh edges
-
-        >>> edges_x = mesh.edges_x
-        >>> edges_y = mesh.edges_y
-        >>> u_ex = -(edges_x[:, 1] / np.sqrt(np.sum(edges_x ** 2, axis=1))) * np.exp(
-        ...     -(edges_x[:, 0] ** 2 + edges_x[:, 1] ** 2) / 6 ** 2
-        ... )
-        >>> u_ey = (edges_y[:, 0] / np.sqrt(np.sum(edges_y ** 2, axis=1))) * np.exp(
-        ...     -(edges_y[:, 0] ** 2 + edges_y[:, 1] ** 2) / 6 ** 2
-        ... )
-        >>> u_e = np.r_[u_ex, u_ey]
-
-        Next, we construct the averaging operator and apply it to
-        the discrete vector quantity to approximate the value at the faces.
-
-        >>> Aef = mesh.average_edge_to_face_vector
-        >>> u_f = Aef @ u_e
-
-        Plot the results,
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(u_e, ax=ax1, v_type="E", view='vec')
-            >>> ax1.set_title("Variable at edges", fontsize=16)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(u_f, ax=ax2, v_type="F", view='vec')
-            >>> ax2.set_title("Averaged to faces", fontsize=16)
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Aef, ms=1)
-            >>> ax1.set_title("Edge Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Face Index", fontsize=12)
-            >>> plt.show()
-        """
-
         if self.dim == 1:
             return self.average_cell_to_face
         elif self.dim == 2:
@@ -4789,89 +3335,6 @@ class DiffOperators(object):
 
     @property
     def average_node_to_cell(self):
-        """Averaging operator from nodes to cell centers (scalar quantities).
-
-        This property constructs a 2nd order averaging operator that maps scalar
-        quantities from nodes to cell centers. This averaging operator is
-        used when a discrete scalar quantity defined on mesh nodes must be
-        projected to cell centers. Once constructed, the operator is stored
-        permanently as a property of the mesh. *See notes*.
-
-        Returns
-        -------
-        (n_cells, n_nodes) scipy.sparse.csr_matrix
-            The scalar averaging operator from nodes to cell centers
-
-        Notes
-        -----
-        Let :math:`\\boldsymbol{\\phi_n}` be a discrete scalar quantity that
-        lives on mesh nodes. **average_node_to_cell** constructs a discrete
-        linear operator :math:`\\mathbf{A_{nc}}` that projects
-        :math:`\\boldsymbol{\\phi_f}` to cell centers, i.e.:
-
-        .. math::
-            \\boldsymbol{\\phi_c} = \\mathbf{A_{nc}} \\, \\boldsymbol{\\phi_n}
-
-        where :math:`\\boldsymbol{\\phi_c}` approximates the value of the scalar
-        quantity at cell centers. For each cell, we are simply averaging
-        the values defined on its nodes. The operation is implemented as a
-        matrix vector product, i.e.::
-
-            phi_c = Anc @ phi_n
-
-        Examples
-        --------
-        Here we compute the values of a scalar function on the nodes. We then create
-        an averaging operator to approximate the function at cell centers. We choose
-        to define a scalar function that is strongly discontinuous in some places to
-        demonstrate how the averaging operator will smooth out discontinuities.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> h = np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Then we Create a scalar variable on nodes
-
-        >>> phi_n = np.zeros(mesh.nN)
-        >>> xy = mesh.nodes
-        >>> phi_n[(xy[:, 1] > 0)] = 25.0
-        >>> phi_n[(xy[:, 1] < -10.0) & (xy[:, 0] > -10.0) & (xy[:, 0] < 10.0)] = 50.0
-
-        Next, we construct the averaging operator and apply it to
-        the discrete scalar quantity to approximate the value at cell centers.
-
-        >>> Anc = mesh.average_node_to_cell
-        >>> phi_c = Anc @ phi_n
-
-        Plot the results,
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(phi_n, ax=ax1, v_type="N")
-            >>> ax1.set_title("Variable at nodes", fontsize=16)
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(phi_c, ax=ax2, v_type="CC")
-            >>> ax2.set_title("Averaged to cell centers", fontsize=16)
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Anc, ms=1)
-            >>> ax1.set_title("Node Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Cell Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_node_to_cell", None) is None:
             # The number of cell centers in each direction
             if self.dim == 1:
@@ -4939,90 +3402,6 @@ class DiffOperators(object):
 
     @property
     def average_node_to_edge(self):
-        """Averaging operator from nodes to edges (scalar quantities).
-
-        This property constructs a 2nd order averaging operator that maps scalar
-        quantities from nodes to edges; scalar at edges is organized in a 1D numpy.array
-        of the form [x-edges, y-edges, z-edges]. This averaging operator is
-        used when a discrete scalar quantity defined on mesh nodes must be
-        projected to edges. Once constructed, the operator is stored
-        permanently as a property of the mesh. *See notes*.
-
-        Returns
-        -------
-        (n_edges, n_nodes) scipy.sparse.csr_matrix
-            The scalar averaging operator from nodes to edges
-
-        Notes
-        -----
-        Let :math:`\\boldsymbol{\\phi_n}` be a discrete scalar quantity that
-        lives on mesh nodes. **average_node_to_edge** constructs a discrete
-        linear operator :math:`\\mathbf{A_{ne}}` that projects
-        :math:`\\boldsymbol{\\phi_n}` to edges, i.e.:
-
-        .. math::
-            \\boldsymbol{\\phi_e} = \\mathbf{A_{ne}} \\, \\boldsymbol{\\phi_n}
-
-        where :math:`\\boldsymbol{\\phi_e}` approximates the value of the scalar
-        quantity at edges. For each edge, we are simply averaging
-        the values defined on the nodes it connects. The operation is implemented as a
-        matrix vector product, i.e.::
-
-            phi_e = Ane @ phi_n
-
-        Examples
-        --------
-        Here we compute the values of a scalar function on the nodes. We then create
-        an averaging operator to approximate the function at the edges. We choose
-        to define a scalar function that is strongly discontinuous in some places to
-        demonstrate how the averaging operator will smooth out discontinuities.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> h = np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Then we create a scalar variable on nodes,
-
-        >>> phi_n = np.zeros(mesh.nN)
-        >>> xy = mesh.nodes
-        >>> phi_n[(xy[:, 1] > 0)] = 25.0
-        >>> phi_n[(xy[:, 1] < -10.0) & (xy[:, 0] > -10.0) & (xy[:, 0] < 10.0)] = 50.0
-
-        Next, we construct the averaging operator and apply it to
-        the discrete scalar quantity to approximate the value on the edges.
-
-        >>> Ane = mesh.average_node_to_edge
-        >>> phi_e = Ane @ phi_n
-
-        Plot the results,
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(phi_n, ax=ax1, v_type="N")
-            >>> ax1.set_title("Variable at nodes")
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(phi_e, ax=ax2, v_type="E")
-            >>> ax2.set_title("Averaged to edges")
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Ane, ms=1)
-            >>> ax1.set_title("Node Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Edge Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_node_to_edge", None) is None:
             # The number of cell centers in each direction
             if self.dim == 1:
@@ -5085,90 +3464,6 @@ class DiffOperators(object):
 
     @property
     def average_node_to_face(self):
-        """Averaging operator from nodes to faces (scalar quantities).
-
-        This property constructs a 2nd order averaging operator that maps scalar
-        quantities from nodes to edges; scalar at faces is organized in a 1D numpy.array
-        of the form [x-faces, y-faces, z-faces]. This averaging operator is
-        used when a discrete scalar quantity defined on mesh nodes must be
-        projected to faces. Once constructed, the operator is stored
-        permanently as a property of the mesh. *See notes*.
-
-        Returns
-        -------
-        (n_faces, n_nodes) scipy.sparse.csr_matrix
-            The scalar averaging operator from nodes to faces
-
-        Notes
-        -----
-        Let :math:`\\boldsymbol{\\phi_n}` be a discrete scalar quantity that
-        lives on mesh nodes. **average_node_to_face** constructs a discrete
-        linear operator :math:`\\mathbf{A_{nf}}` that projects
-        :math:`\\boldsymbol{\\phi_n}` to faces, i.e.:
-
-        .. math::
-            \\boldsymbol{\\phi_f} = \\mathbf{A_{nf}} \\, \\boldsymbol{\\phi_n}
-
-        where :math:`\\boldsymbol{\\phi_f}` approximates the value of the scalar
-        quantity at faces. For each face, we are simply averaging the values at
-        the nodes which outline the face. The operation is implemented as a
-        matrix vector product, i.e.::
-
-            phi_f = Anf @ phi_n
-
-        Examples
-        --------
-        Here we compute the values of a scalar function on the nodes. We then create
-        an averaging operator to approximate the function at the faces. We choose
-        to define a scalar function that is strongly discontinuous in some places to
-        demonstrate how the averaging operator will smooth out discontinuities.
-
-        We start by importing the necessary packages and defining a mesh.
-
-        >>> from discretize import TensorMesh
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> h = np.ones(40)
-        >>> mesh = TensorMesh([h, h], x0="CC")
-
-        Then we, create a scalar variable on nodes
-
-        >>> phi_n = np.zeros(mesh.nN)
-        >>> xy = mesh.nodes
-        >>> phi_n[(xy[:, 1] > 0)] = 25.0
-        >>> phi_n[(xy[:, 1] < -10.0) & (xy[:, 0] > -10.0) & (xy[:, 0] < 10.0)] = 50.0
-
-        Next, we construct the averaging operator and apply it to
-        the discrete scalar quantity to approximate the value on the faces.
-
-        >>> Anf = mesh.average_node_to_face
-        >>> phi_f = Anf @ phi_n
-
-        Plot the results,
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(11, 5))
-            >>> ax1 = fig.add_subplot(121)
-            >>> mesh.plot_image(phi_n, ax=ax1, v_type="N")
-            >>> ax1.set_title("Variable at nodes")
-            >>> ax2 = fig.add_subplot(122)
-            >>> mesh.plot_image(phi_f, ax=ax2, v_type="F")
-            >>> ax2.set_title("Averaged to faces")
-            >>> plt.show()
-
-        Below, we show a spy plot illustrating the sparsity and mapping
-        of the operator
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(9, 9))
-            >>> ax1 = fig.add_subplot(111)
-            >>> ax1.spy(Anf, ms=1)
-            >>> ax1.set_title("Node Index", fontsize=12, pad=5)
-            >>> ax1.set_ylabel("Face Index", fontsize=12)
-            >>> plt.show()
-        """
         if getattr(self, "_average_node_to_face", None) is None:
             # The number of cell centers in each direction
             if self.dim == 1:
@@ -5191,21 +3486,6 @@ class DiffOperators(object):
 
     @property
     def project_face_to_boundary_face(self):
-        """Projection matrix from all faces to boundary faces.
-
-        Constructs and returns a matrix :math:`\\mathbf{P}` that projects from
-        all mesh faces to boundary faces. That is, for a discrete vector
-        :math:`\\mathbf{u}` that lives on the faces, the values on the boundary
-        faces :math:`\\mathbf{u_b}` can be extracted via the following
-        matrix-vector product::
-
-            ub = P @ u
-
-        Returns
-        -------
-        scipy.sparse.csr_matrix
-            (n_boundary_faces, n_faces) Projection matrix with shape
-        """
         # Simple matrix which projects the values of the faces onto the boundary faces
         # Can also be used to "select" the boundary faces
 
@@ -5220,21 +3500,6 @@ class DiffOperators(object):
 
     @property
     def project_edge_to_boundary_edge(self):
-        """Projection matrix from all edges to boundary edges.
-
-        Constructs and returns a matrix :math:`\\mathbf{P}` that projects from
-        all mesh edges to boundary edges. That is, for a discrete vector
-        :math:`\\mathbf{u}` that lives on the edges, the values on the boundary
-        edges :math:`\\mathbf{u_b}` can be extracted via the following
-        matrix-vector product::
-
-            ub = P @ u
-
-        Returns
-        -------
-        (n_boundary_edges, n_edges) scipy.sparse.csr_matrix
-            Projection matrix with shape
-        """
         # Simple matrix which projects the values of the faces onto the boundary faces
         # Can also be used to "select" the boundary faces
 
@@ -5253,21 +3518,6 @@ class DiffOperators(object):
 
     @property
     def project_node_to_boundary_node(self):
-        """Projection matrix from all nodes to boundary nodes.
-
-        Constructs and returns a matrix :math:`\\mathbf{P}` that projects from
-        all mesh nodes to boundary nodes. That is, for a discrete scalar
-        :math:`\\mathbf{u}` that lives on the nodes, the values on the boundary
-        nodes :math:`\\mathbf{u_b}` can be extracted via the following
-        matrix-vector product::
-
-            ub = P @ u
-
-        Returns
-        -------
-        (n_boundary_nodes, n_nodes) scipy.sparse.csr_matrix
-            Projection matrix with shape
-        """
         # Simple matrix which projects the values of the nodes onto the boundary nodes
         # Can also be used to "select" the boundary nodes
 
@@ -5291,13 +3541,6 @@ class DiffOperators(object):
     cellGradz = deprecate_property(
         "cell_gradient_z", "cellGradz", removal_version="1.0.0", future_warn=False
     )
-    nodalGrad = deprecate_property(
-        "nodal_gradient", "nodalGrad", removal_version="1.0.0", future_warn=False
-    )
-    nodalLaplacian = deprecate_property(
-        "nodal_laplacian", "nodalLaplacian", removal_version="1.0.0", future_warn=False
-    )
-    faceDiv = deprecate_property("face_divergence", "faceDiv", removal_version="1.0.0", future_warn=False)
     faceDivx = deprecate_property(
         "face_x_divergence", "faceDivx", removal_version="1.0.0", future_warn=False
     )
@@ -5307,7 +3550,6 @@ class DiffOperators(object):
     faceDivz = deprecate_property(
         "face_z_divergence", "faceDivz", removal_version="1.0.0", future_warn=False
     )
-    edgeCurl = deprecate_property("edge_curl", "edgeCurl", removal_version="1.0.0", future_warn=False)
     _cellGradStencil = deprecate_property(
         "stencil_cell_gradient", "_cellGradStencil", removal_version="1.0.0", future_warn=False
     )
