+++ discretize/operators/inner_products.py
@@ -37,175 +37,6 @@ class InnerProducts(object):
         do_fast=True,
         **kwargs
     ):
-        r"""Generate the face inner product matrix or its inverse.
-
-        This method generates the inner product matrix (or its inverse)
-        when discrete variables are defined on mesh faces. It is also capable of
-        constructing the inner product matrix when physical properties
-        are defined in the form of constitutive relations. For a comprehensive
-        description of the inner product matrices that can be constructed
-        with **get_face_inner_product**, see *Notes*.
-
-        Parameters
-        ----------
-        model : None or numpy.ndarray, optional
-            Parameters defining the material properties for every cell in the mesh.
-            Inner product matrices can be constructed for the following cases:
-
-            - *None* : returns the basic inner product matrix
-            - *(n_cells)* :class:`numpy.ndarray` : returns inner product matrix for an
-              isotropic model. The array contains a scalar physical property value for
-              each cell.
-            - *(n_cells, dim)* :class:`numpy.ndarray` : returns inner product matrix for
-              diagonal anisotropic case. Columns are ordered ``np.c_[σ_xx, σ_yy, σ_zz]``.
-              This can also a be a 1D array with the same number of total elements in
-              column major order.
-            - *(n_cells, 3)* :class:`numpy.ndarray` (``dim`` is 2) or
-              *(n_cells, 6)* :class:`numpy.ndarray` (``dim`` is 3) :
-              returns inner product matrix for full tensor properties case. Columns are
-              ordered ``np.c_[σ_xx, σ_yy, σ_zz, σ_xy, σ_xz, σ_yz]`` This can also be a
-              1D array with the same number of total elements in column major order.
-
-        invert_model : bool, optional
-            The inverse of *model* is used as the physical property.
-        invert_matrix : bool, optional
-            Returns the inverse of the inner product matrix.
-            The inverse not implemented for full tensor properties.
-        do_fast : bool, optional
-            Do a faster implementation (if available).
-
-        Returns
-        -------
-        (n_faces, n_faces) scipy.sparse.csr_matrix
-            inner product matrix
-
-        Notes
-        -----
-        For continuous vector quantities :math:`\vec{u}` and :math:`\vec{w}`
-        whose discrete representations :math:`\mathbf{u}` and :math:`\mathbf{w}`
-        live on the faces, **get_face_inner_product** constructs the inner product matrix
-        :math:`\mathbf{M_\ast}` (or its inverse :math:`\mathbf{M_\ast^{-1}}`) for the
-        following cases:
-
-        **Basic Inner Product:** the inner product between :math:`\vec{u}` and :math:`\vec{w}`
-
-        .. math::
-            \langle \vec{u}, \vec{w} \rangle = \mathbf{u^T \, M \, w}
-
-        **Isotropic Case:** the inner product between :math:`\vec{u}` and :math:`\sigma \vec{w}`
-        where :math:`\sigma` is a scalar function.
-
-        .. math::
-            \langle \vec{u}, \sigma \vec{w} \rangle = \mathbf{u^T \, M_\sigma \, w}
-
-        **Tensor Case:** the inner product between :math:`\vec{u}` and :math:`\Sigma \vec{w}`
-        where :math:`\Sigma` is tensor function; :math:`\sigma_{xy} = \sigma_{xz} = \sigma_{yz} = 0`
-        for diagonal anisotropy.
-
-        .. math::
-            \langle \vec{u}, \Sigma \vec{w} \rangle = \mathbf{u^T \, M_\Sigma \, w}
-            \;\;\; \textrm{where} \;\;\;
-            \Sigma = \begin{bmatrix}
-            \sigma_{xx} & \sigma_{xy} & \sigma_{xz} \\
-            \sigma_{xy} & \sigma_{yy} & \sigma_{yz} \\
-            \sigma_{xz} & \sigma_{yz} & \sigma_{zz}
-            \end{bmatrix}
-
-        Examples
-        --------
-        Here we provide some examples of face inner product matrices.
-        For simplicity, we will work on a 2 x 2 x 2 tensor mesh.
-        As seen below, we begin by constructing and imaging the basic
-        face inner product matrix.
-
-        >>> from discretize import TensorMesh
-        >>> import matplotlib.pyplot as plt
-        >>> import numpy as np
-        >>> import matplotlib as mpl
-
-        >>> h = np.ones(2)
-        >>> mesh = TensorMesh([h, h, h])
-        >>> Mf = mesh.get_face_inner_product()
-
-        >>> fig = plt.figure(figsize=(6, 6))
-        >>> ax = fig.add_subplot(111)
-        >>> ax.imshow(Mf.todense())
-        >>> ax.set_title('Basic Face Inner Product Matrix', fontsize=18)
-        >>> plt.show()
-
-        Next, we consider the case where the physical properties
-        of the cells are defined by consistutive relations. For
-        the isotropic, diagonal anisotropic and full tensor cases,
-        we show the physical property tensor for a single cell.
-
-        Define 4 constitutive parameters and define the tensor
-        for each cell for isotropic, diagonal and tensor cases.
-
-        >>> sig1, sig2, sig3, sig4, sig5, sig6 = 6, 5, 4, 3, 2, 1
-        >>> sig_iso_tensor = sig1 * np.eye(3)
-        >>> sig_diag_tensor = np.diag(np.array([sig1, sig2, sig3]))
-        >>> sig_full_tensor = np.array([
-        ...     [sig1, sig4, sig5],
-        ...     [sig4, sig2, sig6],
-        ...     [sig5, sig6, sig3]
-        ... ])
-
-        Then plot matrix entries,
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(15, 5))
-            >>> ax1 = fig.add_subplot(131)
-            >>> ax1.imshow(sig_iso_tensor)
-            >>> ax1.axis('off')
-            >>> ax1.set_title("Tensor (isotropic)", fontsize=16)
-            >>> ax2 = fig.add_subplot(132)
-            >>> ax2.imshow(sig_diag_tensor)
-            >>> ax2.axis('off')
-            >>> ax2.set_title("Tensor (diagonal anisotropic)", fontsize=16)
-            >>> ax3 = fig.add_subplot(133)
-            >>> ax3.imshow(sig_full_tensor)
-            >>> ax3.axis('off')
-            >>> ax3.set_title("Tensor (full anisotropic)", fontsize=16)
-            >>> plt.show()
-
-        Here, construct and image the face inner product matrices for
-        the isotropic, diagonal anisotropic and full tensor cases.
-        Spy plots are used to demonstrate the sparsity of the inner
-        product matrices.
-
-        Isotropic case:
-
-        >>> v = np.ones(mesh.nC)
-        >>> sig = sig1 * v
-        >>> M1 = mesh.get_face_inner_product(sig)
-
-        Diagonal anisotropic case:
-
-        >>> sig = np.c_[sig1*v, sig2*v, sig3*v]
-        >>> M2 = mesh.get_face_inner_product(sig)
-
-        Full anisotropic case:
-
-        >>> sig = np.tile(np.c_[sig1, sig2, sig3, sig4, sig5, sig6], (mesh.nC, 1))
-        >>> M3 = mesh.get_face_inner_product(sig)
-
-        And then we can plot the sparse representation,
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(12, 4))
-            >>> ax1 = fig.add_subplot(131)
-            >>> ax1.spy(M1, ms=5)
-            >>> ax1.set_title("M (isotropic)", fontsize=16)
-            >>> ax2 = fig.add_subplot(132)
-            >>> ax2.spy(M2, ms=5)
-            >>> ax2.set_title("M (diagonal anisotropic)", fontsize=16)
-            >>> ax3 = fig.add_subplot(133)
-            >>> ax3.spy(M3, ms=5)
-            >>> ax3.set_title("M (full anisotropic)", fontsize=16)
-            >>> plt.show()
-        """
         if "invProp" in kwargs:
             warnings.warn(
                 "The invProp keyword argument has been deprecated, please use invert_model. "
@@ -244,177 +75,6 @@ class InnerProducts(object):
         do_fast=True,
         **kwargs
     ):
-        r"""Generate the edge inner product matrix or its inverse.
-
-        This method generates the inner product matrix (or its inverse)
-        when discrete variables are defined on mesh edges. It is also capable of
-        constructing the inner product matrix when physical properties
-        are defined in the form of constitutive relations. For a comprehensive
-        description of the inner product matrices that can be constructed
-        with **get_edge_inner_product**, see *Notes*.
-
-        Parameters
-        ----------
-        model : None or numpy.ndarray
-            Parameters defining the material properties for every cell in the mesh.
-            Inner product matrices can be constructed for the following cases:
-
-            - *None* : returns the basic inner product matrix
-            - *(n_cells)* :class:`numpy.ndarray` : returns inner product matrix for an
-              isotropic model. The array contains a scalar physical property value for
-              each cell.
-            - *(n_cells, dim)* :class:`numpy.ndarray` : returns inner product matrix for
-              diagonal anisotropic case. Columns are ordered ``np.c_[σ_xx, σ_yy, σ_zz]``.
-              This can also a be a 1D array with the same number of total elements in
-              column major order.
-            - *(n_cells, 3)* :class:`numpy.ndarray` (``dim`` is 2) or
-              *(n_cells, 6)* :class:`numpy.ndarray` (``dim`` is 3) :
-              returns inner product matrix for full tensor properties case. Columns are
-              ordered ``np.c_[σ_xx, σ_yy, σ_zz, σ_xy, σ_xz, σ_yz]`` This can also be a
-              1D array with the same number of total elements in column major order.
-
-        invert_model : bool, optional
-            The inverse of *model* is used as the physical property.
-        invert_matrix : bool, optional
-            Teturns the inverse of the inner product matrix.
-            The inverse not implemented for full tensor properties.
-        do_fast : bool, optional
-            Do a faster implementation (if available).
-
-        Returns
-        -------
-        (n_edges, n_edges) scipy.sparse.csr_matrix
-            inner product matrix
-
-        Notes
-        -----
-        For continuous vector quantities :math:`\vec{u}` and :math:`\vec{w}`
-        whose discrete representations :math:`\mathbf{u}` and :math:`\mathbf{w}`
-        live on the edges, **get_edge_inner_product** constructs the inner product
-        matrix :math:`\mathbf{M_\ast}` (or its inverse :math:`\mathbf{M_\ast^{-1}}`) for
-        the following cases:
-
-        **Basic Inner Product:** the inner product between :math:`\vec{u}` and
-        :math:`\vec{w}`.
-
-        .. math::
-            \langle \vec{u}, \vec{w} \rangle = \mathbf{u^T \, M \, w}
-
-        **Isotropic Case:** the inner product between :math:`\vec{u}` and
-        :math:`\sigma \vec{w}` where :math:`\sigma` is a scalar function.
-
-        .. math::
-            \langle \vec{u}, \sigma \vec{w} \rangle = \mathbf{u^T \, M_\sigma \, w}
-
-        **Tensor Case:** the inner product between :math:`\vec{u}` and
-        :math:`\Sigma \vec{w}` where :math:`\Sigma` is tensor function;
-        :math:`\sigma_{xy} = \sigma_{xz} = \sigma_{yz} = 0` for diagonal anisotropy.
-
-        .. math::
-            \langle \vec{u}, \Sigma \vec{w} \rangle =
-            \mathbf{u^T \, M_\Sigma \, w} \;\;\; \textrm{where} \;\;\;
-            \Sigma = \begin{bmatrix}
-            \sigma_{xx} & \sigma_{xy} & \sigma_{xz} \\
-            \sigma_{xy} & \sigma_{yy} & \sigma_{yz} \\
-            \sigma_{xz} & \sigma_{yz} & \sigma_{zz}
-            \end{bmatrix}
-
-        Examples
-        --------
-        Here we provide some examples of edge inner product matrices.
-        For simplicity, we will work on a 2 x 2 x 2 tensor mesh.
-        As seen below, we begin by constructing and imaging the basic
-        edge inner product matrix.
-
-        >>> from discretize import TensorMesh
-        >>> import matplotlib.pyplot as plt
-        >>> import numpy as np
-        >>> import matplotlib as mpl
-
-        >>> h = np.ones(2)
-        >>> mesh = TensorMesh([h, h, h])
-        >>> Me = mesh.get_edge_inner_product()
-
-        >>> fig = plt.figure(figsize=(6, 6))
-        >>> ax = fig.add_subplot(111)
-        >>> ax.imshow(Me.todense())
-        >>> ax.set_title('Basic Edge Inner Product Matrix', fontsize=18)
-        >>> plt.show()
-
-        Next, we consider the case where the physical properties
-        of the cells are defined by consistutive relations. For
-        the isotropic, diagonal anisotropic and full tensor cases,
-        we show the physical property tensor for a single cell.
-
-        Define 4 constitutive parameters and define the tensor
-        for each cell for isotropic, diagonal and tensor cases.
-
-        >>> sig1, sig2, sig3, sig4, sig5, sig6 = 6, 5, 4, 3, 2, 1
-        >>> sig_iso_tensor = sig1 * np.eye(3)
-        >>> sig_diag_tensor = np.diag(np.array([sig1, sig2, sig3]))
-        >>> sig_full_tensor = np.array([
-        ...     [sig1, sig4, sig5],
-        ...     [sig4, sig2, sig6],
-        ...     [sig5, sig6, sig3]
-        ... ])
-
-        Then plot the matrix entries,
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(15, 5))
-            >>> ax1 = fig.add_subplot(131)
-            >>> ax1.imshow(sig_iso_tensor)
-            >>> ax1.axis('off')
-            >>> ax1.set_title("Tensor (isotropic)", fontsize=16)
-            >>> ax2 = fig.add_subplot(132)
-            >>> ax2.imshow(sig_diag_tensor)
-            >>> ax2.axis('off')
-            >>> ax2.set_title("Tensor (diagonal anisotropic)", fontsize=16)
-            >>> ax3 = fig.add_subplot(133)
-            >>> ax3.imshow(sig_full_tensor)
-            >>> ax3.axis('off')
-            >>> ax3.set_title("Tensor (full anisotropic)", fontsize=16)
-            >>> plt.show()
-
-        Here construct and image the edge inner product matrices for
-        the isotropic, diagonal anisotropic and full tensor cases.
-        Spy plots are used to demonstrate the sparsity of the inner
-        product matrices.
-
-        Isotropic case:
-
-        >>> v = np.ones(mesh.nC)
-        >>> sig = sig1 * v
-        >>> M1 = mesh.get_edge_inner_product(sig)
-
-        Diagonal anisotropic case:
-
-        >>> sig = np.c_[sig1*v, sig2*v, sig3*v]
-        >>> M2 = mesh.get_edge_inner_product(sig)
-
-        Full anisotropic
-
-        >>> sig = np.tile(np.c_[sig1, sig2, sig3, sig4, sig5, sig6], (mesh.nC, 1))
-        >>> M3 = mesh.get_edge_inner_product(sig)
-
-        Then plot the sparse representation,
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(12, 4))
-            >>> ax1 = fig.add_subplot(131)
-            >>> ax1.spy(M1, ms=5)
-            >>> ax1.set_title("M (isotropic)", fontsize=16)
-            >>> ax2 = fig.add_subplot(132)
-            >>> ax2.spy(M2, ms=5)
-            >>> ax2.set_title("M (diagonal anisotropic)", fontsize=16)
-            >>> ax3 = fig.add_subplot(133)
-            >>> ax3.spy(M3, ms=5)
-            >>> ax3.set_title("M (full anisotropic)", fontsize=16)
-            >>> plt.show()
-
-        """
         if "invProp" in kwargs:
             warnings.warn(
                 "The invProp keyword argument has been deprecated, please use invert_model. "
@@ -584,187 +244,6 @@ class InnerProducts(object):
     def get_face_inner_product_deriv(
         self, model, do_fast=True, invert_model=False, invert_matrix=False, **kwargs
     ):
-        r"""Function handle to multiply vector with derivative of face inner product matrix (or its inverse).
-
-        Let :math:`\mathbf{M}(\mathbf{m})` be the face inner product matrix
-        constructed with a set of physical property parameters :math:`\mathbf{m}`
-        (or its inverse). **get_face_inner_product_deriv** constructs a function handle
-
-        .. math::
-            \mathbf{F}(\mathbf{u}) = \mathbf{u}^T \, \frac{\partial \mathbf{M}(\mathbf{m})}{\partial \mathbf{m}}
-
-        which accepts any numpy.array :math:`\mathbf{u}` of shape (n_faces,). That is,
-        **get_face_inner_product_deriv** constructs a function handle for computing
-        the dot product between a vector :math:`\mathbf{u}` and the derivative of the
-        face inner product matrix (or its inverse) with respect to the property parameters.
-        When computed, :math:`\mathbf{F}(\mathbf{u})` returns a ``scipy.sparse.csr_matrix``
-        of shape (n_faces, n_param).
-
-        The function handle can be created for isotropic, diagonal
-        isotropic and full tensor physical properties; see notes.
-
-        Parameters
-        ----------
-        model : numpy.ndarray
-            Parameters defining the material properties for every cell in the mesh.
-            Inner product matrices can be constructed for the following cases:
-
-            - *(n_cells)* :class:`numpy.ndarray` : Isotropic case. *model* contains a
-              scalar physical property value for each cell.
-            - *(n_cells, dim)* :class:`numpy.ndarray` : Diagonal anisotropic case.
-              Columns are ordered ``np.c_[σ_xx, σ_yy, σ_zz]``. This can also a be a 1D
-              array with the same number of total elements in column major order.
-            - *(n_cells, 3)* :class:`numpy.ndarray` (``dim`` is 2) or
-              *(n_cells, 6)* :class:`numpy.ndarray` (``dim`` is 3) : Full tensor properties case. Columns
-              are ordered ``np.c_[σ_xx, σ_yy, σ_zz, σ_xy, σ_xz, σ_yz]`` This can also be
-              a 1D array with the same number of total elements in column major order.
-
-        invert_model : bool, optional
-            The inverse of *model* is used as the physical property.
-        invert_matrix : bool, optional
-            Returns the inverse of the inner product matrix.
-            The inverse not implemented for full tensor properties.
-        do_fast : bool, optional
-            Do a faster implementation (if available).
-
-        Returns
-        -------
-        function
-            The function handle :math:`\mathbf{F}(\mathbf{u})` which accepts a
-            (``n_faces``) :class:`numpy.ndarray` :math:`\mathbf{u}`. The function
-            returns a (``n_faces``, ``n_params``) :class:`scipy.sparse.csr_matrix`.
-
-        Notes
-        -----
-        Let :math:`\mathbf{M}(\mathbf{m})` be the face inner product matrix (or its inverse)
-        for the set of physical property parameters :math:`\mathbf{m}`. And let :math:`\mathbf{u}`
-        be a discrete quantity that lives on the faces. **get_face_inner_product_deriv**
-        creates a function handle for computing the following:
-
-        .. math::
-            \mathbf{F}(\mathbf{u}) = \mathbf{u}^T \, \frac{\partial \mathbf{M}(\mathbf{m})}{\partial \mathbf{m}}
-
-        The dimensions of the sparse matrix constructed by computing :math:`\mathbf{F}(\mathbf{u})`
-        for some :math:`\mathbf{u}` depends on the constitutive relation defined for each cell.
-        These cases are summarized below.
-
-        **Isotropic Case:** The physical property for each cell is defined by a scalar value.
-        Therefore :math:`\mathbf{m}` is a (``n_cells``) :class:`numpy.ndarray`. The sparse matrix
-        output by computing :math:`\mathbf{F}(\mathbf{u})` has shape (``n_faces``, ``n_cells``).
-
-        **Diagonal Anisotropic Case:** In this case, the physical properties for each cell are
-        defined by a diagonal tensor
-
-        .. math::
-            \Sigma = \begin{bmatrix}
-            \sigma_{xx} & 0 & 0 \\
-            0 & \sigma_{yy} & 0 \\
-            0 & 0 & \sigma_{zz}
-            \end{bmatrix}
-
-        Thus there are ``dim * n_cells`` physical property parameters and :math:`\mathbf{m}` is
-        a (``dim * n_cells``) :class:`numpy.ndarray`.  The sparse matrix
-        output by computing :math:`\mathbf{F}(\mathbf{u})` has shape (``n_faces``, ``dim * n_cells``).
-
-        **Full Tensor Case:** In this case, the physical properties for each cell are
-        defined by a full tensor
-
-        .. math::
-            \Sigma = \begin{bmatrix}
-            \sigma_{xx} & \sigma_{xy} & \sigma_{xz} \\
-            \sigma_{xy} & \sigma_{yy} & \sigma_{yz} \\
-            \sigma_{xz} & \sigma_{yz} & \sigma_{zz}
-            \end{bmatrix}
-
-        Thus there are ``6 * n_cells`` physical property parameters in 3 dimensions, or
-        ``3 * n_cells`` physical property parameters in 2 dimensions, and
-        :math:`\mathbf{m}` is a (``n_params``) :class:`numpy.ndarray`.
-        The sparse matrix output by computing :math:`\mathbf{F}(\mathbf{u})`
-        has shape (``n_faces``, ``n_params``).
-
-        Examples
-        --------
-        Here, we construct a 4 cell by 4 cell tensor mesh. For our first example we
-        consider isotropic physical properties; that is, the physical properties
-        of each cell are defined a scalar value. We construct the face inner product
-        matrix and visualize it with a spy plot. We then use
-        **get_face_inner_product_deriv** to construct the function handle
-        :math:`\mathbf{F}(\mathbf{u})` and plot the evaluation
-        of this function on a spy plot.
-
-        >>> from discretize import TensorMesh
-        >>> import matplotlib.pyplot as plt
-        >>> import numpy as np
-        >>> import matplotlib as mpl
-        >>> mpl.rcParams.update({'font.size': 14})
-        >>> np.random.seed(45)
-        >>> mesh = TensorMesh([[(1, 4)], [(1, 4)]])
-
-        Define a model, and a random vector to multiply the derivative with,
-        then we grab the respective derivative function and calculate the
-        sparse matrix,
-
-        >>> m = np.random.rand(mesh.nC)  # physical property parameters
-        >>> u = np.random.rand(mesh.nF)  # vector of shape (n_faces)
-        >>> Mf = mesh.get_face_inner_product(m)
-        >>> F = mesh.get_face_inner_product_deriv(m)  # Function handle
-        >>> dFdm_u = F(u)
-
-        Spy plot for the inner product matrix and its derivative
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(15, 5))
-            >>> ax1 = fig.add_axes([0.05, 0.05, 0.3, 0.85])
-            >>> ax1.spy(Mf, ms=6)
-            >>> ax1.set_title("Face Inner Product Matrix (Isotropic)", fontsize=14, pad=5)
-            >>> ax1.set_xlabel("Face Index", fontsize=12)
-            >>> ax1.set_ylabel("Face Index", fontsize=12)
-            >>> ax2 = fig.add_axes([0.43, 0.05, 0.17, 0.8])
-            >>> ax2.spy(dFdm_u, ms=6)
-            >>> ax2.set_title(
-            ...     "$u^T \, \dfrac{\partial M(m)}{\partial m}$ (Isotropic)",
-            ...     fontsize=14, pad=5
-            ... )
-            >>> ax2.set_xlabel("Parameter Index", fontsize=12)
-            >>> ax2.set_ylabel("Face Index", fontsize=12)
-            >>> plt.show()
-
-        For our second example, the physical properties on the mesh are fully
-        anisotropic; that is, the physical properties of each cell are defined
-        by a tensor with parameters :math:`\sigma_1`, :math:`\sigma_2` and :math:`\sigma_3`.
-        Once again we construct the face inner product matrix and visualize it with a
-        spy plot. We then use **get_face_inner_product_deriv** to construct the
-        function handle :math:`\mathbf{F}(\mathbf{u})` and plot the evaluation
-        of this function on a spy plot.
-
-        >>> m = np.random.rand(mesh.nC, 3)  # anisotropic physical property parameters
-        >>> u = np.random.rand(mesh.nF)     # vector of shape (n_faces)
-        >>> Mf = mesh.get_face_inner_product(m)
-        >>> F = mesh.get_face_inner_product_deriv(m)  # Function handle
-        >>> dFdm_u = F(u)
-
-        Plot the anisotropic inner product matrix and its derivative matrix,
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(15, 5))
-            >>> ax1 = fig.add_axes([0.05, 0.05, 0.3, 0.8])
-            >>> ax1.spy(Mf, ms=6)
-            >>> ax1.set_title("Face Inner Product (Full Tensor)", fontsize=14, pad=5)
-            >>> ax1.set_xlabel("Face Index", fontsize=12)
-            >>> ax1.set_ylabel("Face Index", fontsize=12)
-            >>> ax2 = fig.add_axes([0.4, 0.05, 0.45, 0.85])
-            >>> ax2.spy(dFdm_u, ms=6)
-            >>> ax2.set_title(
-            ...     "$u^T \, \dfrac{\partial M(m)}{\partial m} \;$ (Full Tensor)",
-            ...     fontsize=14, pad=5
-            ... )
-            >>> ax2.set_xlabel("Parameter Index", fontsize=12)
-            >>> ax2.set_ylabel("Face Index", fontsize=12)
-            >>> plt.show()
-
-        """
         if "invProp" in kwargs:
             warnings.warn(
                 "The invProp keyword argument has been deprecated, please use invert_model. "
@@ -797,185 +276,6 @@ class InnerProducts(object):
     def get_edge_inner_product_deriv(
         self, model, do_fast=True, invert_model=False, invert_matrix=False, **kwargs
     ):
-        r"""Function handle to multiply vector with derivative of edge inner product matrix (or its inverse).
-
-        Let :math:`\mathbf{M}(\mathbf{m})` be the edge inner product matrix
-        constructed with a set of physical property parameters :math:`\mathbf{m}`
-        (or its inverse). **get_edge_inner_product_deriv** constructs a function handle
-
-        .. math::
-            \mathbf{F}(\mathbf{u}) = \mathbf{u}^T \, \frac{\partial \mathbf{M}(\mathbf{m})}{\partial \mathbf{m}}
-
-        which accepts any numpy.array :math:`\mathbf{u}` of shape (n_edges,). That is,
-        **get_edge_inner_product_deriv** constructs a function handle for computing
-        the dot product between a vector :math:`\mathbf{u}` and the derivative of the
-        edge inner product matrix (or its inverse) with respect to the property parameters.
-        When computed, :math:`\mathbf{F}(\mathbf{u})` returns a ``scipy.sparse.csr_matrix``
-        of shape (n_edges, n_param).
-
-        The function handle can be created for isotropic, diagonal
-        isotropic and full tensor physical properties; see notes.
-
-        Parameters
-        ----------
-        model : numpy.ndarray
-            Parameters defining the material properties for every cell in the mesh.
-
-            Allows for the following cases:
-
-            - *(n_cells)* :class:`numpy.ndarray` : Isotropic case. *model* contains a
-              scalar physical property value for each cell.
-            - *(n_cells, dim)* :class:`numpy.ndarray` : Diagonal anisotropic case.
-              Columns are ordered ``np.c_[σ_xx, σ_yy, σ_zz]``. This can also a be a 1D
-              array with the same number of total elements in column major order.
-            - *(n_cells, 3)* :class:`numpy.ndarray` (``dim`` is 2) or
-              *(n_cells, 6)* :class:`numpy.ndarray` (``dim`` is 3) : Full tensor properties case. Columns
-              are ordered ``np.c_[σ_xx, σ_yy, σ_zz, σ_xy, σ_xz, σ_yz]`` This can also be
-              a 1D array with the same number of total elements in column major order.
-
-        invert_model : bool, optional
-            The inverse of *model* is used as the physical property.
-        invert_matrix : bool, optional
-            Returns the function handle for the inverse of the inner product matrix
-            The inverse not implemented for full tensor properties.
-        do_fast : bool, optional
-            Do a faster implementation (if available).
-
-        Returns
-        -------
-        function
-            The function handle :math:`\mathbf{F}(\mathbf{u})` which accepts a
-            (``n_edges``) :class:`numpy.ndarray` :math:`\mathbf{u}`. The function
-            returns a (``n_edges``, ``n_params``) :class:`scipy.sparse.csr_matrix`.
-
-        Notes
-        -----
-        Let :math:`\mathbf{M}(\mathbf{m})` be the edge inner product matrix (or its inverse)
-        for the set of physical property parameters :math:`\mathbf{m}`. And let :math:`\mathbf{u}`
-        be a discrete quantity that lives on the edges. **get_edge_inner_product_deriv**
-        creates a function handle for computing the following:
-
-        .. math::
-            \mathbf{F}(\mathbf{u}) = \mathbf{u}^T \, \frac{\partial \mathbf{M}(\mathbf{m})}{\partial \mathbf{m}}
-
-        The dimensions of the sparse matrix constructed by computing :math:`\mathbf{F}(\mathbf{u})`
-        for some :math:`\mathbf{u}` depends on the constitutive relation defined for each cell.
-        These cases are summarized below.
-
-        **Isotropic Case:** The physical property for each cell is defined by a scalar value.
-        Therefore :math:`\mathbf{m}` is a (``n_cells``) :class:`numpy.ndarray`. The sparse matrix
-        output by computing :math:`\mathbf{F}(\mathbf{u})` has shape (``n_edges``, ``n_cells``).
-
-        **Diagonal Anisotropic Case:** In this case, the physical properties for each cell are
-        defined by a diagonal tensor
-
-        .. math::
-            \Sigma = \begin{bmatrix}
-            \sigma_{xx} & 0 & 0 \\
-            0 & \sigma_{yy} & 0 \\
-            0 & 0 & \sigma_{zz}
-            \end{bmatrix}
-
-        Thus there are ``dim * n_cells`` physical property parameters and :math:`\mathbf{m}` is
-        a (``dim * n_cells``) :class:`numpy.ndarray`.  The sparse matrix
-        output by computing :math:`\mathbf{F}(\mathbf{u})` has shape (``n_edges``, ``dim * n_cells``).
-
-        **Full Tensor Case:** In this case, the physical properties for each cell are
-        defined by a full tensor
-
-        .. math::
-            \Sigma = \begin{bmatrix}
-            \sigma_{xx} & \sigma_{xy} & \sigma_{xz} \\
-            \sigma_{xy} & \sigma_{yy} & \sigma_{yz} \\
-            \sigma_{xz} & \sigma_{yz} & \sigma_{zz}
-            \end{bmatrix}
-
-        Thus there are ``6 * n_cells`` physical property parameters in 3 dimensions, or
-        ``3 * n_cells`` physical property parameters in 2 dimensions, and :math:`\mathbf{m}` is
-        a (``n_params``) :class:`numpy.ndarray`. The sparse matrix
-        output by computing :math:`\mathbf{F}(\mathbf{u})` has shape (``n_edges``, ``n_params``).
-
-        Examples
-        --------
-        Here, we construct a 4 cell by 4 cell tensor mesh. For our first example we
-        consider isotropic physical properties; that is, the physical properties
-        of each cell are defined a scalar value. We construct the edge inner product
-        matrix and visualize it with a spy plot. We then use
-        **get_edge_inner_product_deriv** to construct the function handle
-        :math:`\mathbf{F}(\mathbf{u})` and plot the evaluation of this function on a spy
-        plot.
-
-        >>> from discretize import TensorMesh
-        >>> import matplotlib.pyplot as plt
-        >>> import numpy as np
-        >>> import matplotlib as mpl
-        >>> mpl.rcParams.update({'font.size': 14})
-        >>> np.random.seed(45)
-        >>> mesh = TensorMesh([[(1, 4)], [(1, 4)]])
-
-        Next we create a random isotropic model vector, and a random vector to multiply
-        the derivative with (for illustration purposes).
-
-        >>> m = np.random.rand(mesh.nC)  # physical property parameters
-        >>> u = np.random.rand(mesh.nF)  # vector of shape (n_edges)
-        >>> Me = mesh.get_edge_inner_product(m)
-        >>> F = mesh.get_edge_inner_product_deriv(m)  # Function handle
-        >>> dFdm_u = F(u)
-
-        Plot inner product matrix and its derivative matrix
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(15, 5))
-            >>> ax1 = fig.add_axes([0.05, 0.05, 0.3, 0.8])
-            >>> ax1.spy(Me, ms=6)
-            >>> ax1.set_title("Edge Inner Product Matrix (Isotropic)", fontsize=14, pad=5)
-            >>> ax1.set_xlabel("Edge Index", fontsize=12)
-            >>> ax1.set_ylabel("Edge Index", fontsize=12)
-            >>> ax2 = fig.add_axes([0.43, 0.05, 0.17, 0.8])
-            >>> ax2.spy(dFdm_u, ms=6)
-            >>> ax2.set_title(
-            ...     "$u^T \, \dfrac{\partial M(m)}{\partial m}$ (Isotropic)",
-            ...     fontsize=14, pad=5
-            ... )
-            >>> ax2.set_xlabel("Parameter Index", fontsize=12)
-            >>> ax2.set_ylabel("Edge Index", fontsize=12)
-            >>> plt.show()
-
-        For our second example, the physical properties on the mesh are fully
-        anisotropic; that is, the physical properties of each cell are defined
-        by a tensor with parameters :math:`\sigma_1`, :math:`\sigma_2` and :math:`\sigma_3`.
-        Once again we construct the edge inner product matrix and visualize it with a
-        spy plot. We then use **get_edge_inner_product_deriv** to construct the
-        function handle :math:`\mathbf{F}(\mathbf{u})` and plot the evaluation
-        of this function on a spy plot.
-
-        >>> m = np.random.rand(mesh.nC, 3)  # physical property parameters
-        >>> u = np.random.rand(mesh.nF)     # vector of shape (n_edges)
-        >>> Me = mesh.get_edge_inner_product(m)
-        >>> F = mesh.get_edge_inner_product_deriv(m)  # Function handle
-        >>> dFdm_u = F(u)
-
-        Plot the anisotropic inner product matrix and its derivative matrix
-
-        .. collapse:: Expand to show scripting for plot
-
-            >>> fig = plt.figure(figsize=(15, 5))
-            >>> ax1 = fig.add_axes([0.05, 0.05, 0.3, 0.8])
-            >>> ax1.spy(Me, ms=6)
-            >>> ax1.set_title("Edge Inner Product (Full Tensor)", fontsize=14, pad=5)
-            >>> ax1.set_xlabel("Edge Index", fontsize=12)
-            >>> ax1.set_ylabel("Edge Index", fontsize=12)
-            >>> ax2 = fig.add_axes([0.4, 0.05, 0.45, 0.8])
-            >>> ax2.spy(dFdm_u, ms=6)
-            >>> ax2.set_title(
-            ...     "$u^T \, \dfrac{\partial M(m)}{\partial m} \;$ (Full Tensor)",
-            ...     fontsize=14, pad=5
-            ... )
-            >>> ax2.set_xlabel("Parameter Index", fontsize=12)
-            >>> ax2.set_ylabel("Edge Index", fontsize=12)
-            >>> plt.show()
-        """
         if "invProp" in kwargs:
             warnings.warn(
                 "The invProp keyword argument has been deprecated, please use invert_model. "
@@ -1494,23 +794,3 @@ class InnerProducts(object):
             return PXXX
 
         return Pxxx
-
-    # DEPRECATED
-    getFaceInnerProduct = deprecate_method(
-        "get_face_inner_product", "getFaceInnerProduct", removal_version="1.0.0", future_warn=False
-    )
-    getEdgeInnerProduct = deprecate_method(
-        "get_edge_inner_product", "getEdgeInnerProduct", removal_version="1.0.0", future_warn=False
-    )
-    getFaceInnerProductDeriv = deprecate_method(
-        "get_face_inner_product_deriv",
-        "getFaceInnerProductDeriv",
-        removal_version="1.0.0",
-        future_warn=False
-    )
-    getEdgeInnerProductDeriv = deprecate_method(
-        "get_edge_inner_product_deriv",
-        "getEdgeInnerProductDeriv",
-        removal_version="1.0.0",
-        future_warn=False
-    )
